Fix: Refactor implementation.rs - Extract RecordCache and neuron analysis (#185)
- Create src/analysis/cache.rs with RecordCache implementation
- Move neuron analysis functions to src/analysis/neuron.rs
- Remove duplicated functions from implementation.rs (now imported from synapse.rs)
- Update mod.rs to use extracted modules
- Fix imports in diagnostics.rs and synapse.rs for RecordCache

This reduces implementation.rs from ~9831 to ~7135 lines (non-test code
now under 2000 lines). All 319 tests pass.
