+++ src/analysis/implementation.rs
@@ -1,27 +1,15 @@
 use crate::types::DiscoverRecord;
-use crate::{
-    AnalyzeNeuronsInput, AnalyzeSynapsesInput, CandidateNeuronJson, CandidateSynapseJson,
-    SynapseJson,
-};
-use anyhow::{anyhow, Context, Result};
+use crate::{AnalyzeSynapsesInput, CandidateSynapseJson, SynapseJson};
+use anyhow::{anyhow, Result};
 
 // Import shared types from the new module structure
-use crate::analysis::shared::{AnalyzeNeuronsResult, AnalyzeSynapsesResult, TimingScope};
+use crate::analysis::shared::{AnalyzeSynapsesResult, TimingScope};
 
 // Import activation functions from the dedicated activation module (Issue #266, #238)
-// Note: get_bias_range, get_bias_values moved to weights module usage (Issue #270)
-// Note: Target simulation functions moved here from implementation.rs (Issue #238)
+// Note: Activation functions and specs moved to neuron.rs for neuron analysis (Issue #185)
 #[cfg(test)]
 use crate::analysis::activation::can_use_hard_tanh;
-use crate::analysis::activation::{
-    absolute_activation, activation_name_to_gpu_id, arctan_activation, bent_identity_activation,
-    bipolar_activation, clipped_activation, elu_activation, gelu_activation,
-    get_target_simulation_fn, get_target_simulation_mode, hard_tanh_activation,
-    has_sufficient_output_variance, identity_activation, is_threshold_activation,
-    logistic_activation, mish_activation, relu6_activation, softplus_activation,
-    softsign_activation, tanh_activation, ActivationCandidateSpec, TargetSimulationMode,
-    ACTIVATION_SPECS,
-};
+use crate::analysis::activation::get_target_simulation_fn;
 
 // Import memory and platform utilities from dedicated modules (Issue #267)
 // Note: cap_gpu_batch_size_by_bytes, check_system_memory_requirements, detect_memory_tier,
@@ -29,7 +17,7 @@ use crate::analysis::activation::{
 // DEFAULT_GPU_BATCH_SIZE, HIGH_PERF_GPU_BATCH_SIZE, LOW_MEMORY_GPU_BATCH_SIZE moved to
 // gpu/analyzer.rs (Issue #273)
 // get_work_queue_capacity moved to gpu/queue.rs usage (Issue #274)
-use crate::analysis::utils::{check_memory_for_parquet, verbose_enabled};
+use crate::analysis::utils::verbose_enabled;
 
 // Import deadline handling and logging utilities from dedicated module (Issue #268)
 #[cfg(test)]
@@ -41,35 +29,26 @@ use crate::analysis::utils::{
 };
 
 // Import sample data structures from dedicated module (Issue #269)
-// Note: Many GPU buffer types moved to gpu/analyzer.rs (Issue #273)
-// HarmfulStats, HelpfulStats moved to gpu/queue.rs usage (Issue #274)
-// HelpfulStats still needed for tests - see test module imports
+// Note: compute_source_variance_discount and more moved to neuron.rs (Issue #185)
 use crate::analysis::samples::{
-    compute_source_variance_discount, constant_source_effect_threshold_from_env, HelpfulSample,
-    NeuronStats, EPSILON,
+    constant_source_effect_threshold_from_env, HelpfulSample, NeuronStats, EPSILON,
 };
 
 // Import weight calculation functions from dedicated module (Issue #270)
 use crate::analysis::weights::{
-    calculate_optimal_bias, calculate_optimal_identity_outgoing_and_bias,
     calculate_optimal_outgoing_weight, clamp_weight_update_delta,
-    coordinated_structural_activation_delta, MAX_OUTGOING_WEIGHT,
+    coordinated_structural_activation_delta,
 };
 
 // Import diagnostics and rejection tracking from dedicated module (Issue #271)
+// Note: NeuronDiagnostics, FocusTargetFilterResult, etc. moved to neuron.rs (Issue #185)
 use crate::analysis::diagnostics::{
-    compute_impact_scores_for_discounting, filter_focus_targets_for_neuron_analysis,
-    require_unique_focus, FocusTargetFilterResult, NeuronDiagnostics, TargetDiagnostics, TargetMap,
+    compute_impact_scores_for_discounting, require_unique_focus, TargetDiagnostics, TargetMap,
     ThresholdContext,
 };
 
 // Import GPU infrastructure from dedicated modules (Issue #272, #273, #274)
-// Most GPU types and functions are now used only via GpuAnalyzer/GpuWorkQueue
-// GPU_INIT_TIMEOUT_SECS now used by gpu/queue.rs (Issue #274)
-use crate::analysis::gpu::{GpuAnalyzer, GpuEvaluator, GpuWorkQueue};
-
-// Re-import shared types needed by implementation
-use crate::analysis::shared::{NeuronNoCandidateReason, NeuronNoCandidateSummary};
+use crate::analysis::gpu::{GpuAnalyzer, GpuWorkQueue};
 
 // Import test-specific types from diagnostics and shared
 #[cfg(test)]
@@ -79,7 +58,7 @@ use crate::analysis::shared::SynapseNoCandidateReason;
 
 // bytemuck::Zeroable moved to gpu/analyzer.rs (Issue #273)
 // crossbeam_channel::{bounded, Receiver, Sender} moved to gpu/queue.rs (Issue #274)
-use once_cell::sync::OnceCell;
+// once_cell::sync::OnceCell moved to cache.rs (Issue #185)
 use rayon::prelude::*;
 use std::cmp::Ordering;
 use std::collections::{HashMap, HashSet};
@@ -123,14 +102,16 @@ const MIN_NEURON_SAMPLE_COUNT: usize = 10;
 // Types moved to shared.rs - using imports from there
 // OrderedNeuron moved to utils/deadline.rs (Issue #268)
 
-type RecordCacheLoader = dyn Fn(&str, &str) -> Result<Vec<DiscoverRecord>> + Send + Sync + 'static;
-type CachedNeuronRecords = OnceCell<Arc<Vec<DiscoverRecord>>>;
+// RecordCache moved to cache.rs (Issue #185)
+pub(crate) use super::cache::RecordCache;
 
-pub(crate) struct RecordCache {
-    parquet_file: String,
-    pub(crate) cache: Mutex<HashMap<String, Arc<CachedNeuronRecords>>>,
-    loader: Arc<RecordCacheLoader>,
-}
+// Import shared helper functions from synapse module (Issue #185)
+// These are used by synapse analysis (neuron analysis is in neuron.rs)
+use super::synapse::{
+    build_ordered_neurons, build_samples_for_locality_group, compute_synapse_improvement_and_count,
+    group_sources_by_locality, truncate_combined_synapse_candidate_sets,
+    MIN_GROUP_SIZE_FOR_LOCALITY,
+};
 
 // RecordCacheProvider, compute_impact_scores_for_discounting, RejectionReason
 // moved to crate::analysis::diagnostics module (Issue #271)
@@ -141,134 +122,7 @@ pub(crate) struct RecordCache {
 // NeuronRejectionDetail, NeuronDiagnosticEntry, NeuronDiagnostics
 // moved to crate::analysis::diagnostics module (Issue #271)
 
-impl RecordCache {
-    /// Create a cache that automatically chooses the best loading strategy based on
-    /// available system memory:
-    ///
-    /// - **Pre-loaded mode** (fast): Loads entire parquet file into memory upfront.
-    ///   Used when there's sufficient RAM (3× file size + 1GB headroom).
-    ///
-    /// - **Lazy-loaded mode** (memory-efficient): Loads records on-demand per neuron.
-    ///   Slower (O(N) parquet scans for N neurons) but works on memory-constrained systems.
-    ///
-    /// This ensures discovery works on any modern Mac/PC, adapting to available resources.
-    pub(crate) fn new_adaptive(parquet_file: &str) -> Result<Self> {
-        // Check if we have enough memory for pre-loading
-        match check_memory_for_parquet(parquet_file) {
-            Ok(()) => {
-                // Sufficient memory - use fast pre-loaded mode
-                Self::new_preloaded_internal(parquet_file)
-            }
-            Err(memory_error) => {
-                // Insufficient memory - fall back to lazy loading
-                eprintln!(
-                    "[NEAT-AI-Discovery] Insufficient memory for pre-loading. \
-                     Falling back to lazy-loading mode (slower but memory-efficient)."
-                );
-                if verbose_enabled() {
-                    eprintln!("[NEAT-AI-Discovery][verbose] Memory check failed: {memory_error}");
-                }
-                Self::new_lazy(parquet_file)
-            }
-        }
-    }
-
-    /// Create a lazy-loading cache that loads records on-demand.
-    /// Slower than pre-loaded mode but uses minimal memory.
-    fn new_lazy(parquet_file: &str) -> Result<Self> {
-        use crate::parquet_format::read_records_from_parquet;
-
-        eprintln!(
-            "[NEAT-AI-Discovery] Using lazy-loading mode for parquet file. \
-             This is slower but uses less memory."
-        );
-
-        Ok(Self {
-            parquet_file: parquet_file.to_string(),
-            cache: Mutex::new(HashMap::new()),
-            loader: Arc::new(move |file: &str, neuron_uuid: &str| {
-                read_records_from_parquet(file, neuron_uuid)
-            }),
-        })
-    }
-
-    /// Internal pre-loaded implementation (called when memory check passes).
-    fn new_preloaded_internal(parquet_file: &str) -> Result<Self> {
-        use crate::parquet_format::read_all_records_grouped_by_neuron;
-        use std::time::Instant;
-
-        let start = Instant::now();
-        let grouped_records = read_all_records_grouped_by_neuron(parquet_file)
-            .with_context(|| format!("Failed to pre-load records from {parquet_file}"))?;
-
-        let neuron_count = grouped_records.len();
-        let total_records: usize = grouped_records.values().map(|v| v.len()).sum();
-
-        // Pre-populate the cache with all loaded records
-        let mut cache_map: HashMap<String, Arc<CachedNeuronRecords>> = HashMap::new();
-        for (uuid, mut records) in grouped_records {
-            records.sort_by_key(|r| r.obs_index);
-            let cell = OnceCell::new();
-            let _ = cell.set(Arc::new(records));
-            cache_map.insert(uuid, Arc::new(cell));
-        }
-
-        let elapsed = start.elapsed();
-        if verbose_enabled() {
-            eprintln!(
-                "[NEAT-AI-Discovery][verbose] Pre-loaded {} neurons with {} total records from parquet in {:.2}s",
-                neuron_count,
-                total_records,
-                elapsed.as_secs_f64()
-            );
-        }
-
-        Ok(Self {
-            parquet_file: parquet_file.to_string(),
-            cache: Mutex::new(cache_map),
-            loader: Arc::new(|_file: &str, _neuron_uuid: &str| {
-                // This loader should never be called for pre-loaded cache
-                Ok(Vec::new())
-            }),
-        })
-    }
-
-    #[cfg(test)]
-    fn with_loader(parquet_file: &str, loader: Arc<RecordCacheLoader>) -> Self {
-        Self {
-            parquet_file: parquet_file.to_string(),
-            cache: Mutex::new(HashMap::new()),
-            loader,
-        }
-    }
-
-    pub(crate) fn get(&self, neuron_uuid: &str) -> Result<Arc<Vec<DiscoverRecord>>> {
-        let cache_entry = {
-            let mut cache = self.cache.lock().expect("record cache mutex poisoned");
-            cache
-                .entry(neuron_uuid.to_string())
-                .or_insert_with(|| Arc::new(OnceCell::new()))
-                .clone()
-        };
-
-        let loader = Arc::clone(&self.loader);
-        let parquet_file = self.parquet_file.clone();
-        let context_uuid = neuron_uuid.to_string();
-        let load_uuid = context_uuid.clone();
-
-        let arc_records = cache_entry
-            .get_or_try_init(move || -> Result<Arc<Vec<DiscoverRecord>>> {
-                let mut records = loader(&parquet_file, &load_uuid)?;
-                records.sort_by_key(|record| record.obs_index);
-                Ok(Arc::new(records))
-            })
-            .with_context(|| {
-                format!("Failed to read discovery records for neuron {context_uuid}")
-            })?;
-
-        Ok(Arc::clone(arc_records))
-    }
-}
+// RecordCache implementation moved to cache.rs (Issue #185)
 
 // require_unique_focus moved to crate::analysis::diagnostics module (Issue #271)
 
@@ -323,2556 +177,6 @@ impl RecordCache {
 // - merge_batch_results()
 
 // Shader constants moved to gpu/analyzer.rs (Issue #273)
-
-fn build_ordered_neurons(creature: &crate::CreatureJson) -> Vec<OrderedNeuron> {
-    let mut ordered = Vec::with_capacity(creature.input + creature.neurons.len());
-
-    for input_index in 0..creature.input {
-        ordered.push(OrderedNeuron {
-            uuid: format!("input-{input_index}"),
-            index: input_index,
-        });
-    }
-
-    for (offset, neuron) in creature.neurons.iter().enumerate() {
-        ordered.push(OrderedNeuron {
-            uuid: neuron.uuid.clone(),
-            index: creature.input + offset,
-        });
-    }
-
-    ordered
-}
-
-// TargetData, TargetMap moved to crate::analysis::diagnostics module (Issue #271)
-
-// ============================================================================
-// Issue #221: Sample Locality Grouping
-// ============================================================================
-//
-// When analysing multiple source neurons for the same target, sources that
-// share the same obs_indices can benefit from batched sample building.
-//
-// The grouping algorithm:
-// 1. Compute a "locality key" for each source based on its obs_indices
-// 2. Group sources with identical (or highly overlapping) locality keys
-// 3. Build samples for each group in a single pass through the target data
-
-/// Minimum number of sources in a group to make shared sample building worthwhile.
-/// Below this threshold, the overhead of grouping exceeds the benefit.
-const MIN_GROUP_SIZE_FOR_LOCALITY: usize = 3;
-
-/// Minimum overlap fraction required to group sources together.
-/// Sources are grouped if they share at least this fraction of their obs_indices.
-const MIN_LOCALITY_OVERLAP: f32 = 0.8;
-
-/// Represents a group of sources with similar obs_index coverage.
-/// Sources in the same group can share sample building overhead.
-struct SampleLocalityGroup<'a> {
-    /// The source neurons in this group
-    sources: Vec<(&'a OrderedNeuron, Arc<Vec<DiscoverRecord>>)>,
-    /// Representative obs_indices for this group (from the first source)
-    _representative_indices: HashSet<u32>,
-}
-
-/// Extract obs_indices from source records.
-fn extract_obs_indices(records: &[DiscoverRecord]) -> HashSet<u32> {
-    records
-        .iter()
-        .filter(|r| r.activation.is_finite())
-        .map(|r| r.obs_index)
-        .collect()
-}
-
-/// Compute the overlap fraction between two sets of obs_indices.
-/// Returns a value in [0, 1] representing what fraction of the smaller set
-/// is contained in the larger set.
-fn compute_obs_index_overlap(a: &HashSet<u32>, b: &HashSet<u32>) -> f32 {
-    if a.is_empty() || b.is_empty() {
-        return 0.0;
-    }
-    let intersection_size = a.intersection(b).count();
-    let min_size = a.len().min(b.len());
-    intersection_size as f32 / min_size as f32
-}
-
-/// Group sources by sample locality for efficient batch processing.
-///
-/// Sources with high obs_index overlap (≥80%) are grouped together so that
-/// sample building can be done in a single pass through the target data
-/// rather than N separate passes.
-///
-/// # Issue #221: Sample Locality for Correlated Source Neurons
-///
-/// Expected benefits for typical creatures:
-/// - 100 sources, same obs_indices: 1 group (100x reduction in target lookups)
-/// - 100 sources, 80% overlap: ~5 groups (20x reduction)
-/// - 100 sources, no overlap: 100 groups (no change)
-fn group_sources_by_locality<'a>(
-    sources: &[(&'a OrderedNeuron, Arc<Vec<DiscoverRecord>>)],
-) -> Vec<SampleLocalityGroup<'a>> {
-    if sources.len() < MIN_GROUP_SIZE_FOR_LOCALITY {
-        // Not enough sources to benefit from grouping
-        return sources
-            .iter()
-            .map(|(neuron, records)| {
-                let indices = extract_obs_indices(records);
-                SampleLocalityGroup {
-                    sources: vec![(neuron, Arc::clone(records))],
-                    _representative_indices: indices,
-                }
-            })
-            .collect();
-    }
-
-    // Extract obs_indices for each source (done once, reused for grouping)
-    let source_indices: Vec<HashSet<u32>> = sources
-        .iter()
-        .map(|(_, records)| extract_obs_indices(records))
-        .collect();
-
-    let mut groups: Vec<SampleLocalityGroup<'a>> = Vec::new();
-    let mut assigned: Vec<bool> = vec![false; sources.len()];
-
-    for i in 0..sources.len() {
-        if assigned[i] {
-            continue;
-        }
-
-        let my_indices = &source_indices[i];
-        if my_indices.is_empty() {
-            // Source has no valid indices - put it in its own group
-            assigned[i] = true;
-            groups.push(SampleLocalityGroup {
-                sources: vec![(sources[i].0, Arc::clone(&sources[i].1))],
-                _representative_indices: my_indices.clone(),
-            });
-            continue;
-        }
-
-        // Start a new group with this source
-        let mut group_sources = vec![(sources[i].0, Arc::clone(&sources[i].1))];
-        assigned[i] = true;
-
-        // Find all other sources with high overlap
-        for j in (i + 1)..sources.len() {
-            if assigned[j] {
-                continue;
-            }
-
-            let other_indices = &source_indices[j];
-            let overlap = compute_obs_index_overlap(my_indices, other_indices);
-
-            if overlap >= MIN_LOCALITY_OVERLAP {
-                group_sources.push((sources[j].0, Arc::clone(&sources[j].1)));
-                assigned[j] = true;
-            }
-        }
-
-        groups.push(SampleLocalityGroup {
-            sources: group_sources,
-            _representative_indices: my_indices.clone(),
-        });
-    }
-
-    groups
-}
-
-/// Build samples for all sources in a locality group efficiently.
-///
-/// This uses `TargetMap::build_samples_for_group` to build samples for all
-/// sources in a single pass through the target data.
-fn build_samples_for_locality_group(
-    group: &SampleLocalityGroup<'_>,
-    target_map: &TargetMap,
-) -> Vec<(String, Vec<HelpfulSample>, usize)> {
-    if group.sources.len() == 1 {
-        // Single source - use standard path (no overhead)
-        let (source, records) = &group.sources[0];
-        let samples = target_map.build_samples_from(records);
-        return vec![(source.uuid.clone(), samples, records.len())];
-    }
-
-    // Multiple sources - use batched sample building
-    let source_records: Vec<(&str, &[DiscoverRecord])> = group
-        .sources
-        .iter()
-        .map(|(source, records)| (source.uuid.as_str(), records.as_slice()))
-        .collect();
-
-    let samples_batch = target_map.build_samples_for_group(&source_records);
-
-    group
-        .sources
-        .iter()
-        .zip(samples_batch)
-        .map(|((source, records), samples)| (source.uuid.clone(), samples, records.len()))
-        .collect()
-}
-
-/// Compute expected gain for a coordinated "replace synapse with neuron" group.
-///
-/// This models the coordinated operation sequence:
-/// 1) remove direct synapse (source -> target)
-/// 2) add hidden neuron with (source -> newNeuron) and (newNeuron -> target)
-///
-/// We approximate the effect of synapse removal in the VALUE domain by adjusting
-/// the target error per-sample:
-///   error' = error + old_weight * source_activation
-///
-/// This is fast and deterministic, and is sufficient for IDENTITY targets (the
-/// common case for our coordinated-structural regression tests).
-///
-/// Notes (7-Jan-2026):
-/// - This intentionally uses the linear model (no saturation-aware simulation),
-///   because we do not have a reliable way to reconstruct `target_value` and
-///   `target_activation` after removing the existing synapse.
-#[allow(clippy::too_many_arguments)]
-pub(crate) fn expected_gain_replace_synapse_with_hidden_neuron(
-    cache: &RecordCache,
-    source_uuid: &str,
-    target_uuid: &str,
-    old_weight: f32,
-    incoming_weight: f32,
-    outgoing_weight: f32,
-    bias: f32,
-    squash: &str,
-) -> Option<f32> {
-    let from_records_arc = cache.get(source_uuid).ok()?;
-    let target_records_arc = cache.get(target_uuid).ok()?;
-    if from_records_arc.is_empty() || target_records_arc.is_empty() {
-        return None;
-    }
-
-    let target_map = TargetMap::from_records(target_records_arc.as_ref());
-    if target_map.map.is_empty() {
-        return None;
-    }
-
-    let mut samples = target_map.build_samples_from(from_records_arc.as_ref());
-    if samples.is_empty() {
-        return None;
-    }
-
-    // Adjust baseline errors for the removal of the existing direct synapse.
-    for s in &mut samples {
-        s.avg_error += old_weight * s.activation;
-    }
-
-    let total_baseline_error_sq: f32 = samples.iter().map(|s| s.avg_error * s.avg_error).sum();
-    if total_baseline_error_sq <= EPSILON {
-        return None;
-    }
-
-    if squash.eq_ignore_ascii_case("ReLU") {
-        let (improvement, _improved, _total) = compute_relu_improvement_and_count(
-            samples.as_slice(),
-            incoming_weight,
-            outgoing_weight,
-            bias,
-            total_baseline_error_sq,
-            None, // linear domain (see docstring)
-        );
-        return Some(improvement);
-    }
-
-    // Map squash names to activation functions used by add-neurons discovery.
-    let activation_fn: fn(f32) -> f32 = match squash {
-        "GELU" => gelu_activation,
-        "ELU" => elu_activation,
-        "Softplus" => softplus_activation,
-        "LOGISTIC" => logistic_activation,
-        "TANH" => tanh_activation,
-        "IDENTITY" => identity_activation,
-        "BIPOLAR" => bipolar_activation,
-        "CLIPPED" => clipped_activation,
-        "ABSOLUTE" => absolute_activation,
-        "Mish" => mish_activation,
-        "HARD_TANH" => hard_tanh_activation,
-        "Softsign" => softsign_activation,
-        "BentIdentity" => bent_identity_activation,
-        "Arctan" => arctan_activation,
-        "ReLU6" => relu6_activation,
-        _ => return None,
-    };
-
-    let (improvement, _improved, _total) = compute_activation_improvement_and_count(
-        samples.as_slice(),
-        incoming_weight,
-        outgoing_weight,
-        bias,
-        activation_fn,
-        total_baseline_error_sq,
-        None, // linear domain (see docstring)
-    );
-    Some(improvement)
-}
-
-/// Deterministic UUID generator for coordinated structural `addNeuron` operations.
-///
-/// This deliberately avoids Rust's `Hash` because the default hasher is not stable
-/// across processes. We use a simple FNV-1a 64-bit hash and return a short hex
-/// suffix so candidates are replayable and readable.
-pub(crate) fn deterministic_coordinated_neuron_uuid(
-    source_uuid: &str,
-    target_uuid: &str,
-    squash: &str,
-    incoming_weight: f32,
-    outgoing_weight: f32,
-    bias: f32,
-) -> String {
-    // 7-Jan-2026: keep stable across runs, but include weights/bias so distinct ops don't collide.
-    let key = format!(
-        "replace-synapse-with-neuron|{source_uuid}|{target_uuid}|{squash}|{incoming_weight:.6}|{outgoing_weight:.6}|{bias:.6}"
-    );
-    // FNV-1a 64-bit
-    let mut hash: u64 = 0xcbf29ce484222325;
-    for b in key.as_bytes() {
-        hash ^= *b as u64;
-        hash = hash.wrapping_mul(0x100000001b3);
-    }
-    format!("coordinated-hidden-{hash:016x}")
-}
-
-/// Build samples for testing. In production, use TargetMap::from_records() and
-/// TargetMap::build_samples_from() for better performance when processing
-/// multiple sources against the same target.
-#[cfg(test)]
-fn build_samples(
-    target_records: &[DiscoverRecord],
-    from_records: &[DiscoverRecord],
-) -> Vec<HelpfulSample> {
-    if target_records.is_empty() || from_records.is_empty() {
-        return Vec::new();
-    }
-
-    // Build map from obs_index to target data (error, value, activation)
-    let target_map = TargetMap::from_records(target_records);
-
-    if target_map.map.is_empty() {
-        return Vec::new();
-    }
-
-    target_map.build_samples_from(from_records)
-}
-
-/// Computes the sign of a weight as an i8 for use in the candidate key.
-/// Returns 1 for positive weights, -1 for negative, and 0 for zero (though this
-/// shouldn't happen in practice).
-fn weight_sign(weight: f32) -> i8 {
-    if weight > 0.0 {
-        1
-    } else if weight < 0.0 {
-        -1
-    } else {
-        0
-    }
-}
-
-/// Returns true if this add-neuron candidate is "extreme" enough to warrant a conservative pair.
-///
-/// We intentionally base this on incoming weight and bias (not outgoing), because outgoing
-/// weight is already clamped and ReLU candidates commonly use outgoing=0.1 by design.
-fn upsert_candidate(
-    map: &mut HashMap<(String, String, String, i8, i8), CandidateNeuronJson>,
-    candidate: CandidateNeuronJson,
-) {
-    use std::collections::hash_map::Entry;
-
-    // Key includes signs of BOTH incoming_weight AND outgoing_weight so that:
-    // 1. Different ReLU orientations (incoming_weight ±1) are kept separately
-    // 2. Split-error complementary pairs (same incoming_weight, opposite outgoing_weight)
-    //    are also kept separately - one pushes output UP, one pushes DOWN
-    let key = (
-        candidate.source_neuron_uuid.clone(),
-        candidate.target_neuron_uuid.clone(),
-        candidate.squash.clone(),
-        weight_sign(candidate.incoming_weight),
-        weight_sign(candidate.outgoing_weight),
-    );
-
-    match map.entry(key) {
-        Entry::Occupied(mut entry) => {
-            // Issue #128: Compare by expected creature score gain
-            if candidate.expected_creature_score_gain > entry.get().expected_creature_score_gain {
-                entry.insert(candidate);
-            }
-        }
-        Entry::Vacant(entry) => {
-            entry.insert(candidate);
-        }
-    }
-}
-
-/// Result from ReLU evaluation (split by target error sign)
-struct SplitReluResult {
-    /// Candidate for samples with positive error (output should be higher)
-    positive_error_candidate: Option<CandidateNeuronJson>,
-    /// Candidate for samples with negative error (output should be lower)
-    negative_error_candidate: Option<CandidateNeuronJson>,
-}
-
-// hard_tanh function moved to crate::analysis::weights module (Issue #270)
-// get_target_activation_fn, get_target_simulation_fn, get_target_simulation_mode,
-// TargetSimulationMode, can_use_hard_tanh moved to crate::analysis::activation (Issue #238)
-
-/// Combined computation of improvement and count for ReLU candidates.
-/// Single pass over samples for better cache efficiency.
-///
-/// When `target_activation_fn` is Some, simulates the target neuron's actual activation
-/// function for more accurate improvement estimates. Otherwise falls back to linear approximation.
-///
-/// IMPORTANT: The `bias` parameter is critical for accurate predictions. It shifts the ReLU
-/// activation threshold, affecting which samples produce non-zero output. When bias > 0,
-/// more samples activate; when bias < 0, fewer samples activate. Excluding bias causes
-/// significant prediction errors.
-///
-/// CRITICAL DOMAIN FIX (v0.1.120): When using target_activation_fn simulation,
-/// both baseline and new error must be computed in ACTIVATION domain.
-///
-/// Returns (improvement_percentage, improved_count, total_count)
-fn compute_relu_improvement_and_count(
-    samples: &[HelpfulSample],
-    incoming_weight: f32,
-    outgoing_weight: f32,
-    bias: f32,
-    total_baseline_error_sq: f32,
-    target_activation_fn: Option<fn(f32) -> f32>,
-) -> (f32, u32, u32) {
-    if total_baseline_error_sq <= EPSILON || samples.is_empty() {
-        return (0.0, 0, samples.len() as u32);
-    }
-
-    let mut baseline_error_sq_sum = 0.0f32; // ACTIVATION domain when simulating
-    let mut new_error_sq_sum = 0.0f32;
-    let mut improved_count = 0u32;
-    let mut worsened_count = 0u32;
-
-    for sample in samples.iter() {
-        let pre_activation = incoming_weight * sample.activation + bias;
-        let relu_output = pre_activation.max(0.0);
-        let contribution = outgoing_weight * relu_output;
-
-        let (baseline_error, new_error) = if let Some(target_fn) = target_activation_fn {
-            // CRITICAL: Use ACTIVATION domain for BOTH baseline and new error.
-            let target_value = unsafe { sample.target_value.unwrap_unchecked() };
-            let target_activation = unsafe { sample.target_activation.unwrap_unchecked() };
-            let desired_value = target_value + sample.avg_error;
-            let expected = target_fn(desired_value);
-
-            // Baseline error in ACTIVATION domain
-            let baseline_err = expected - target_activation;
-
-            // New error in ACTIVATION domain
-            let new_input = target_value + contribution;
-            let new_err = expected - target_fn(new_input);
-
-            (baseline_err, new_err)
-        } else {
-            // Linear approximation: both errors in VALUE domain
-            let baseline_err = sample.avg_error;
-            let new_err = sample.avg_error - contribution;
-
-            (baseline_err, new_err)
-        };
-
-        if baseline_error.is_finite() {
-            baseline_error_sq_sum += baseline_error * baseline_error;
-        }
-        if new_error.is_finite() {
-            new_error_sq_sum += new_error * new_error;
-        }
-
-        // Sample is improved if |new_error| < |baseline_error| (consistent domain)
-        if new_error.abs() + EPSILON < baseline_error.abs() {
-            improved_count += 1;
-        } else if new_error.abs() > baseline_error.abs() + EPSILON {
-            worsened_count += 1;
-        }
-    }
-
-    // Use computed ACTIVATION domain baseline when simulating, else use passed VALUE domain
-    let effective_baseline = if target_activation_fn.is_some() {
-        baseline_error_sq_sum
-    } else {
-        total_baseline_error_sq
-    };
-
-    let improvement = if effective_baseline > EPSILON {
-        (effective_baseline - new_error_sq_sum) / effective_baseline
-    } else {
-        0.0
-    };
-    let improvement = if improvement.is_finite() {
-        improvement
-    } else {
-        0.0
-    };
-
-    let total_count = samples.len() as u32;
-    let _ = worsened_count; // Kept for potential future use
-    (improvement, improved_count, total_count)
-}
-
-/// Combined computation of improvement and count for activation candidates.
-/// Single pass over samples for better cache efficiency.
-///
-/// When `target_activation_fn` is Some, simulates the target neuron's actual activation
-/// function for more accurate improvement estimates. Otherwise falls back to linear approximation.
-///
-/// CRITICAL DOMAIN FIX (v0.1.120): When using target_activation_fn simulation,
-/// both baseline and new error must be computed in ACTIVATION domain.
-///
-/// Returns (improvement_percentage, improved_count, total_count)
-fn compute_activation_improvement_and_count(
-    samples: &[HelpfulSample],
-    incoming_weight: f32,
-    outgoing_weight: f32,
-    bias: f32,
-    activation_fn: fn(f32) -> f32,
-    total_baseline_error_sq: f32,
-    target_activation_fn: Option<fn(f32) -> f32>,
-) -> (f32, u32, u32) {
-    if total_baseline_error_sq <= EPSILON || samples.is_empty() {
-        return (0.0, 0, samples.len() as u32);
-    }
-
-    let mut baseline_error_sq_sum = 0.0f32; // ACTIVATION domain when simulating
-    let mut new_error_sq_sum = 0.0f32;
-    let mut improved_count = 0u32;
-    let total_count = samples.len() as u32;
-
-    for sample in samples {
-        let pre_activation = incoming_weight * sample.activation + bias;
-        let neuron_output = activation_fn(pre_activation);
-        let contribution = outgoing_weight * neuron_output;
-
-        let (baseline_error, new_error) = if let Some(target_fn) = target_activation_fn {
-            // CRITICAL: Use ACTIVATION domain for BOTH baseline and new error.
-            let target_value = unsafe { sample.target_value.unwrap_unchecked() };
-            let target_activation = unsafe { sample.target_activation.unwrap_unchecked() };
-            let desired_value = target_value + sample.avg_error;
-            let expected = target_fn(desired_value);
-
-            // Baseline error in ACTIVATION domain
-            let baseline_err = expected - target_activation;
-
-            // New error in ACTIVATION domain
-            let new_input = target_value + contribution;
-            let new_err = expected - target_fn(new_input);
-
-            (baseline_err, new_err)
-        } else {
-            // Linear approximation: both errors in VALUE domain
-            (sample.avg_error, sample.avg_error - contribution)
-        };
-
-        if baseline_error.is_finite() {
-            baseline_error_sq_sum += baseline_error * baseline_error;
-        }
-        if new_error.is_finite() {
-            new_error_sq_sum += new_error * new_error;
-        }
-
-        // Sample is improved if |new_error| < |baseline_error| (consistent domain)
-        if new_error.abs() + EPSILON < baseline_error.abs() {
-            improved_count += 1;
-        }
-    }
-
-    // Use computed ACTIVATION domain baseline when simulating, else use passed VALUE domain
-    let effective_baseline = if target_activation_fn.is_some() {
-        baseline_error_sq_sum
-    } else {
-        total_baseline_error_sq
-    };
-
-    let improvement = if effective_baseline > EPSILON {
-        (effective_baseline - new_error_sq_sum) / effective_baseline
-    } else {
-        0.0
-    };
-    let improvement = if improvement.is_finite() {
-        improvement
-    } else {
-        0.0
-    };
-
-    (improvement, improved_count, total_count)
-}
-
-/// Wrapper for tests - computes improvement only.
-/// NOTE: For ReLU candidates, bias affects which samples activate. Pass the actual bias
-/// that will be used with the new neuron for accurate predictions.
-#[cfg(test)]
-fn compute_net_improvement_with_squash(
-    samples: &[HelpfulSample],
-    incoming_weight: f32,
-    outgoing_weight: f32,
-    bias: f32,
-    total_baseline_error_sq: f32,
-    target_squash: Option<&str>,
-) -> f32 {
-    let target_activation_fn = get_target_simulation_fn(samples, target_squash);
-    let (improvement, _, _) = compute_relu_improvement_and_count(
-        samples,
-        incoming_weight,
-        outgoing_weight,
-        bias,
-        total_baseline_error_sq,
-        target_activation_fn,
-    );
-    improvement
-}
-
-/// Compute synapse improvement accounting for target neuron's activation function.
-///
-/// For direct synapse connections (source → target), the contribution is `weight × source_activation`.
-/// This function simulates the target's activation function to predict accurate improvement,
-/// avoiding overprediction near saturation for HARD_TANH, TANH, LOGISTIC, etc.
-///
-/// Returns improvement_percentage only. Used in tests; production uses compute_synapse_improvement_and_count.
-#[cfg(test)]
-fn compute_synapse_improvement_with_target_squash(
-    samples: &[HelpfulSample],
-    weight: f32,
-    total_baseline_error_sq: f32,
-    target_squash: Option<&str>,
-) -> f32 {
-    if total_baseline_error_sq <= EPSILON || samples.is_empty() {
-        return 0.0;
-    }
-
-    let target_sim = get_target_simulation_mode(samples, target_squash);
-
-    let mut baseline_error_sq_sum = 0.0f32; // ACTIVATION domain when simulating
-    let mut new_error_sq_sum = 0.0f32;
-
-    for sample in samples {
-        // Direct synapse contribution: weight × source_activation
-        let contribution = weight * sample.activation;
-
-        let new_error = match target_sim {
-            TargetSimulationMode::None => {
-                // Linear approximation - assumes contribution directly reduces error (VALUE domain).
-                sample.avg_error - contribution
-            }
-            TargetSimulationMode::Full(target_fn) => {
-                // Saturation-aware model (ACTIVATION domain).
-                let target_value = sample.target_value.unwrap();
-                let target_activation = sample.target_activation.unwrap();
-                let desired_value = target_value + sample.avg_error;
-                let expected = target_fn(desired_value);
-
-                let baseline_err = expected - target_activation;
-                if baseline_err.is_finite() {
-                    baseline_error_sq_sum += baseline_err * baseline_err;
-                }
-
-                let new_input = target_value + contribution;
-                expected - target_fn(new_input)
-            }
-            TargetSimulationMode::ApproximateValueFromActivation(target_fn) => {
-                // Saturation-aware model (ACTIVATION domain), approximating missing target_value.
-                let target_activation = sample.target_activation.unwrap();
-                let target_value = sample.target_value.unwrap_or(target_activation);
-                let desired_value = target_value + sample.avg_error;
-                let expected = target_fn(desired_value);
-
-                let baseline_err = expected - target_activation;
-                if baseline_err.is_finite() {
-                    baseline_error_sq_sum += baseline_err * baseline_err;
-                }
-
-                let new_input = target_value + contribution;
-                expected - target_fn(new_input)
-            }
-        };
-
-        if new_error.is_finite() {
-            new_error_sq_sum += new_error * new_error;
-        }
-    }
-
-    let effective_baseline = match target_sim {
-        TargetSimulationMode::None => total_baseline_error_sq,
-        _ => baseline_error_sq_sum,
-    };
-    if effective_baseline <= EPSILON {
-        return 0.0;
-    }
-
-    let improvement = (effective_baseline - new_error_sq_sum) / effective_baseline;
-    if improvement.is_finite() {
-        improvement
-    } else {
-        0.0
-    }
-}
-
-/// Compute improvement, improved count, and worsened count for synapse candidates.
-/// All counts use the same saturation-aware methodology for consistency.
-///
-/// CRITICAL DOMAIN FIX (v0.1.120): When using target_activation_fn simulation,
-/// both baseline and new error must be computed in ACTIVATION domain. The passed-in
-/// total_baseline_error_sq is in VALUE domain, so we compute our own ACTIVATION
-/// domain baseline when simulating.
-///
-/// Returns (improvement_percentage, improved_count, worsened_count, total_count)
-fn compute_synapse_improvement_and_count(
-    samples: &[HelpfulSample],
-    weight: f32,
-    total_baseline_error_sq: f32,
-    target_squash: Option<&str>,
-) -> (f32, u32, u32, u32) {
-    if total_baseline_error_sq <= EPSILON || samples.is_empty() {
-        return (0.0, 0, 0, samples.len() as u32);
-    }
-
-    let target_sim = get_target_simulation_mode(samples, target_squash);
-
-    let mut baseline_error_sq_sum = 0.0f32; // ACTIVATION domain when simulating
-    let mut new_error_sq_sum = 0.0f32;
-    let mut improved_count = 0u32;
-    let mut worsened_count = 0u32;
-    let total_count = samples.len() as u32;
-
-    for sample in samples {
-        let contribution = weight * sample.activation;
-
-        let (baseline_error, new_error) = match target_sim {
-            TargetSimulationMode::None => {
-                // Linear approximation: both errors in VALUE domain.
-                (sample.avg_error, sample.avg_error - contribution)
-            }
-            TargetSimulationMode::Full(target_fn) => {
-                // CRITICAL: Use ACTIVATION domain for BOTH baseline and new error.
-                // avg_error is in VALUE domain, but MSE is measured in ACTIVATION domain.
-                let target_value = sample.target_value.unwrap();
-                let target_activation = sample.target_activation.unwrap();
-                let desired_value = target_value + sample.avg_error;
-                let expected = target_fn(desired_value);
-
-                let baseline_err = expected - target_activation;
-                let new_input = target_value + contribution;
-                let new_err = expected - target_fn(new_input);
-
-                (baseline_err, new_err)
-            }
-            TargetSimulationMode::ApproximateValueFromActivation(target_fn) => {
-                // As above, but approximate missing target_value from the observed activation.
-                let target_activation = sample.target_activation.unwrap();
-                let target_value = sample.target_value.unwrap_or(target_activation);
-                let desired_value = target_value + sample.avg_error;
-                let expected = target_fn(desired_value);
-
-                let baseline_err = expected - target_activation;
-                let new_input = target_value + contribution;
-                let new_err = expected - target_fn(new_input);
-
-                (baseline_err, new_err)
-            }
-        };
-
-        if baseline_error.is_finite() {
-            baseline_error_sq_sum += baseline_error * baseline_error;
-        }
-        if new_error.is_finite() {
-            new_error_sq_sum += new_error * new_error;
-        }
-
-        // Count improved/worsened samples using consistent domain comparison
-        if new_error.abs() + EPSILON < baseline_error.abs() {
-            improved_count += 1;
-        } else if new_error.abs() > baseline_error.abs() + EPSILON {
-            worsened_count += 1;
-        }
-    }
-
-    // Use computed ACTIVATION domain baseline when simulating, else use passed VALUE domain.
-    let effective_baseline = match target_sim {
-        TargetSimulationMode::None => total_baseline_error_sq,
-        _ => baseline_error_sq_sum,
-    };
-
-    let improvement = if effective_baseline > EPSILON {
-        (effective_baseline - new_error_sq_sum) / effective_baseline
-    } else {
-        0.0
-    };
-    let improvement = if improvement.is_finite() {
-        improvement
-    } else {
-        0.0
-    };
-
-    (improvement, improved_count, worsened_count, total_count)
-}
-
-/// Wrapper for tests - counts improved samples only.
-#[cfg(test)]
-fn count_improved_samples(
-    samples: &[HelpfulSample],
-    incoming_weight: f32,
-    outgoing_weight: f32,
-    bias: f32,
-    target_squash: Option<&str>,
-) -> (u32, u32) {
-    let total_baseline_error_sq: f32 = samples.iter().map(|s| s.avg_error * s.avg_error).sum();
-    let target_activation_fn = get_target_simulation_fn(samples, target_squash);
-    let (_, improved, total) = compute_relu_improvement_and_count(
-        samples,
-        incoming_weight,
-        outgoing_weight,
-        bias,
-        total_baseline_error_sq,
-        target_activation_fn,
-    );
-    (improved, total)
-}
-
-/// Evaluate ReLU candidates by splitting samples based on TARGET neuron's error sign.
-///
-/// This is the PRIMARY approach for ReLU evaluation. It finds candidates for both directions:
-/// - **Positive-error samples** (output should be HIGHER): compute weight that pushes UP
-/// - **Negative-error samples** (output should be LOWER): compute weight that pushes DOWN
-///
-/// For each direction:
-/// 1. Compute optimal weight from the error subset
-/// 2. Evaluate NET improvement across ALL samples
-/// 3. Return candidate if it passes threshold
-///
-/// This is the correct approach for directional activations like ReLU because:
-/// - ReLU can only push output in ONE direction (based on outgoing weight sign)
-/// - Averaging over all samples cancels out when errors are split ~50/50
-/// - We evaluate source activations as-is (we don't care how they were calculated)
-fn evaluate_relu_candidates_split<G: GpuEvaluator>(
-    gpu: &G,
-    source_uuid: &str,
-    target_uuid: &str,
-    samples: &[HelpfulSample],
-    threshold: f32,
-    target_squash: Option<&str>,
-) -> Result<SplitReluResult> {
-    // Split samples by error sign
-    let positive_error_samples: Vec<HelpfulSample> = samples
-        .iter()
-        .filter(|s| s.avg_error > EPSILON)
-        .copied()
-        .collect();
-
-    let negative_error_samples: Vec<HelpfulSample> = samples
-        .iter()
-        .filter(|s| s.avg_error < -EPSILON)
-        .copied()
-        .collect();
-
-    let mut result = SplitReluResult {
-        positive_error_candidate: None,
-        negative_error_candidate: None,
-    };
-
-    // Compute total baseline error across ALL samples (for net improvement calculation)
-    let total_baseline_error_sq: f32 = samples.iter().map(|s| s.avg_error * s.avg_error).sum();
-
-    if total_baseline_error_sq <= EPSILON {
-        return Ok(result);
-    }
-
-    // Get target activation function for accurate simulation (ReLU, HARD_TANH, etc.)
-    let target_activation_fn = get_target_simulation_fn(samples, target_squash);
-
-    // For positive errors (output should be higher), compute optimal weight from subset
-    // then evaluate the NET effect across ALL samples.
-    // We evaluate BOTH ReLU orientations (positive and negative incoming weight) and pick best.
-    if positive_error_samples.len() >= MIN_NEURON_SAMPLE_COUNT {
-        let (positive_stats, negative_stats, pos_baseline_error_sq) =
-            gpu.evaluate_relu(&positive_error_samples, threshold)?;
-
-        // Try both orientations and pick the best
-        let orientations = [positive_stats, negative_stats];
-        let mut best_candidate: Option<CandidateNeuronJson> = None;
-        let mut best_improvement = threshold;
-
-        for stats in orientations {
-            if let Some(mut candidate) = stats.evaluate(
-                source_uuid,
-                target_uuid,
-                threshold,
-                pos_baseline_error_sq,
-                &positive_error_samples,
-            ) {
-                // Compute net improvement across ALL samples (single pass)
-                // CRITICAL: Include candidate.bias for accurate prediction
-                let (net_improvement, improved, total) = compute_relu_improvement_and_count(
-                    samples,
-                    candidate.incoming_weight,
-                    candidate.outgoing_weight,
-                    candidate.bias,
-                    total_baseline_error_sq,
-                    target_activation_fn,
-                );
-                candidate.improved_count = improved;
-                candidate.total_count = total;
-
-                // Issue #128: Update creature-level metrics
-                if net_improvement > best_improvement {
-                    candidate.expected_creature_error_reduction = net_improvement;
-                    candidate.expected_creature_score_gain = net_improvement;
-                    best_improvement = net_improvement;
-                    best_candidate = Some(candidate);
-                }
-            }
-        }
-        result.positive_error_candidate = best_candidate;
-    }
-
-    // For negative errors (output should be lower).
-    // We evaluate BOTH ReLU orientations (positive and negative incoming weight) and pick best.
-    if negative_error_samples.len() >= MIN_NEURON_SAMPLE_COUNT {
-        let (positive_stats, negative_stats, neg_baseline_error_sq) =
-            gpu.evaluate_relu(&negative_error_samples, threshold)?;
-
-        // Try both orientations and pick the best
-        let orientations = [positive_stats, negative_stats];
-        let mut best_candidate: Option<CandidateNeuronJson> = None;
-        let mut best_improvement = threshold;
-
-        for stats in orientations {
-            if let Some(mut candidate) = stats.evaluate(
-                source_uuid,
-                target_uuid,
-                threshold,
-                neg_baseline_error_sq,
-                &negative_error_samples,
-            ) {
-                // Compute net improvement across ALL samples (single pass)
-                // CRITICAL: Include candidate.bias for accurate prediction
-                let (net_improvement, improved, total) = compute_relu_improvement_and_count(
-                    samples,
-                    candidate.incoming_weight,
-                    candidate.outgoing_weight,
-                    candidate.bias,
-                    total_baseline_error_sq,
-                    target_activation_fn,
-                );
-                candidate.improved_count = improved;
-                candidate.total_count = total;
-
-                // Issue #128: Update creature-level metrics
-                if net_improvement > best_improvement {
-                    candidate.expected_creature_error_reduction = net_improvement;
-                    candidate.expected_creature_score_gain = net_improvement;
-                    best_improvement = net_improvement;
-                    best_candidate = Some(candidate);
-                }
-            }
-        }
-        result.negative_error_candidate = best_candidate;
-    }
-
-    Ok(result)
-}
-
-/// Helper for split-error evaluation: compute optimal weight from subset, evaluate on all samples.
-///
-/// This is the core of the split-error fix for non-ReLU activations. By computing
-/// the optimal weight from a specific error subset (positive or negative), we get
-/// a weight that's tuned to help that subset. We then evaluate the NET improvement
-/// across ALL samples to ensure the candidate doesn't hurt the other subset more
-/// than it helps the target subset.
-#[allow(clippy::too_many_arguments)]
-fn evaluate_activation_for_subset<G: GpuEvaluator>(
-    gpu: &G,
-    source_uuid: &str,
-    target_uuid: &str,
-    subset_samples: &[HelpfulSample], // Used to compute optimal weight
-    all_samples: &[HelpfulSample],    // Used to compute net improvement
-    spec: &ActivationCandidateSpec,
-    target_squash: Option<&str>,
-    total_baseline_error_sq: f32,
-    target_activation_fn: Option<fn(f32) -> f32>,
-) -> Result<Option<CandidateNeuronJson>> {
-    if subset_samples.len() < MIN_NEURON_SAMPLE_COUNT {
-        return Ok(None);
-    }
-
-    let activation_type = activation_name_to_gpu_id(spec.name);
-
-    let mut best_candidate: Option<CandidateNeuronJson> = None;
-    // v0.1.136: Fixed threshold bug - use 0.0 instead of threshold.
-    // The calling code in evaluate_activation_candidate handles threshold vs fallback
-    // logic. If we initialise to threshold here, candidates with 0 < improvement <= threshold
-    // are silently dropped, breaking the fallback mechanism for split-error evaluation.
-    let mut best_net_improvement = 0.0;
-
-    for &orientation in spec.orientations {
-        for &scale in spec.scales {
-            let incoming_weight = orientation * scale;
-
-            // Compute optimal weight from SUBSET samples using GPU
-            let (sum_activation_sq, sum_error_activation) = match gpu.evaluate_activation(
-                subset_samples,
-                activation_type,
-                orientation,
-                scale,
-            ) {
-                Ok(result) => (result.0, result.1),
-                Err(_) => {
-                    // Fall back to CPU on GPU error
-                    let mut sum_act_sq = 0.0;
-                    let mut sum_err_act = 0.0;
-                    for sample in subset_samples {
-                        let pre_activation = incoming_weight * sample.activation;
-                        let output = (spec.activation)(pre_activation);
-                        if output.is_finite() {
-                            sum_act_sq += output * output;
-                            sum_err_act += output * sample.avg_error;
-                        }
-                    }
-                    (sum_act_sq, sum_err_act)
-                }
-            };
-
-            let (outgoing_weight, optimal_bias) = if spec.name == "IDENTITY" {
-                match calculate_optimal_identity_outgoing_and_bias(subset_samples, incoming_weight)
-                {
-                    Some((w, b)) => (w, b),
-                    None => continue,
-                }
-            } else {
-                // Use shared weight calculation with ratio validation
-                let outgoing_weight = match calculate_optimal_outgoing_weight(
-                    sum_error_activation,
-                    sum_activation_sq,
-                    incoming_weight,
-                ) {
-                    Some(w) => w,
-                    None => continue, // Skip if weight is invalid or ratio too small
-                };
-
-                // Calculate optimal bias from subset
-                let optimal_bias = calculate_optimal_bias(
-                    subset_samples,
-                    incoming_weight,
-                    outgoing_weight,
-                    spec.activation,
-                    spec.name,
-                    None,
-                    target_squash,
-                );
-                (outgoing_weight, optimal_bias)
-            };
-
-            // Issue #123: Check for saturation - reject if neuron output is nearly constant.
-            // Large bias values (e.g., 5.0 for SOFTSIGN) can cause the neuron to saturate,
-            // producing nearly identical output regardless of input. Such neurons cannot
-            // reduce error correlation and lead to massive prediction failures.
-            if !has_sufficient_output_variance(
-                all_samples,
-                incoming_weight,
-                optimal_bias,
-                spec.activation,
-            ) {
-                continue;
-            }
-
-            // CRITICAL: Evaluate NET improvement across ALL samples
-            // This ensures the candidate helps the target subset more than it hurts the other
-            let (net_improvement, improved_count, total_count) =
-                compute_activation_improvement_and_count(
-                    all_samples,
-                    incoming_weight,
-                    outgoing_weight,
-                    optimal_bias,
-                    spec.activation,
-                    total_baseline_error_sq,
-                    target_activation_fn,
-                );
-
-            // Only consider candidates with positive NET improvement
-            if net_improvement <= 0.0 {
-                continue;
-            }
-
-            // Apply validity filters
-            let absolute_improvement = net_improvement * total_baseline_error_sq;
-            if absolute_improvement < 0.001 {
-                continue;
-            }
-
-            if spec.name == "IDENTITY" && optimal_bias.abs() < 0.01 {
-                continue;
-            }
-
-            // Track best candidate
-            if net_improvement > best_net_improvement {
-                // Guard rail: do not return candidates with absurd bias values.
-                let bias_abs_max =
-                    crate::analysis::utils::sensible_bias_abs_max_for_squash(spec.name);
-                if !optimal_bias.is_finite() || optimal_bias.abs() > bias_abs_max {
-                    continue;
-                }
-
-                best_net_improvement = net_improvement;
-
-                let target_stats = NeuronStats::from_samples(all_samples).map(|s| s.to_json());
-                // Issue #128: Use creature-level metrics
-                best_candidate = Some(CandidateNeuronJson {
-                    source_neuron_uuid: source_uuid.to_string(),
-                    target_neuron_uuid: target_uuid.to_string(),
-                    source_neuron_index: None, // Set during impact discounting
-                    target_neuron_index: None, // Set during impact discounting
-                    incoming_weight,
-                    outgoing_weight,
-                    squash: spec.name.to_string(),
-                    bias: optimal_bias,
-                    comment: None,
-                    target_neuron_impact: 1.0,
-                    expected_creature_error_reduction: net_improvement,
-                    expected_creature_score_gain: net_improvement,
-                    improved_count,
-                    total_count,
-                    target_neuron_stats: target_stats,
-                });
-            }
-        }
-    }
-
-    Ok(best_candidate)
-}
-
-fn evaluate_activation_candidate<G: GpuEvaluator>(
-    gpu: &G,
-    source_uuid: &str,
-    target_uuid: &str,
-    samples: &[HelpfulSample],
-    threshold: f32,
-    spec: &ActivationCandidateSpec,
-    target_squash: Option<&str>,
-) -> Result<Option<CandidateNeuronJson>> {
-    if samples.len() < MIN_NEURON_SAMPLE_COUNT {
-        return Ok(None);
-    }
-
-    let activation_type = activation_name_to_gpu_id(spec.name);
-
-    let mut best_candidate: Option<CandidateNeuronJson> = None;
-    let mut best_score = threshold;
-    let mut fallback_candidate: Option<CandidateNeuronJson> = None;
-    let mut fallback_score = f32::MIN;
-
-    // v0.1.135: Split-error evaluation for all activations (not just ReLU).
-    // When errors are split ~50/50 between positive and negative, computing
-    // optimal weight from ALL samples averages to near-zero, giving weak
-    // predictions that are often wrong in practice.
-    //
-    // The fix: split samples by error sign, compute optimal weight from each
-    // subset, then evaluate NET improvement across ALL samples.
-    let positive_error_samples: Vec<HelpfulSample> = samples
-        .iter()
-        .filter(|s| s.avg_error > EPSILON)
-        .copied()
-        .collect();
-
-    let negative_error_samples: Vec<HelpfulSample> = samples
-        .iter()
-        .filter(|s| s.avg_error < -EPSILON)
-        .copied()
-        .collect();
-
-    // Compute total_baseline_error_sq across ALL samples (for net improvement)
-    let mut total_baseline_error_sq = 0.0;
-    for sample in samples {
-        if sample.avg_error.is_finite() {
-            total_baseline_error_sq += sample.avg_error * sample.avg_error;
-        }
-    }
-
-    // Get target activation function for net improvement calculation
-    let target_activation_fn = get_target_simulation_fn(samples, target_squash);
-
-    // v0.1.136: Track whether split-error evaluation was properly attempted.
-    // If BOTH subsets had enough samples but NEITHER produced candidates,
-    // the errors are truly split and there's no reliable weight direction.
-    // In this case, we should NOT fall back to all-samples evaluation,
-    // which would produce unreliable small-improvement predictions.
-    let positive_subset_valid = positive_error_samples.len() >= MIN_NEURON_SAMPLE_COUNT;
-    let negative_subset_valid = negative_error_samples.len() >= MIN_NEURON_SAMPLE_COUNT;
-    let split_error_attempted = positive_subset_valid && negative_subset_valid;
-
-    // Evaluate candidates from BOTH error subsets
-    // This ensures we find the best direction even with split errors
-    for error_samples in [&positive_error_samples, &negative_error_samples] {
-        if error_samples.len() < MIN_NEURON_SAMPLE_COUNT {
-            continue;
-        }
-
-        // Compute baseline for this subset (used for weight calculation)
-        let subset_baseline_sq: f32 = error_samples
-            .iter()
-            .map(|s| s.avg_error * s.avg_error)
-            .sum();
-
-        if subset_baseline_sq <= EPSILON {
-            continue;
-        }
-
-        if let Some(candidate) = evaluate_activation_for_subset(
-            gpu,
-            source_uuid,
-            target_uuid,
-            error_samples, // Compute weight from subset
-            samples,       // Evaluate improvement on ALL samples
-            spec,
-            target_squash,
-            total_baseline_error_sq,
-            target_activation_fn,
-        )? {
-            // Track best and fallback candidates from split evaluation
-            // Issue #128: Use expected_creature_score_gain for comparison
-            if candidate.expected_creature_score_gain > best_score {
-                best_score = candidate.expected_creature_score_gain;
-                best_candidate = Some(candidate.clone());
-            }
-            if candidate.expected_creature_score_gain > fallback_score
-                && candidate.expected_creature_score_gain > 0.0
-            {
-                fallback_score = candidate.expected_creature_score_gain;
-                fallback_candidate = Some(candidate);
-            }
-        }
-    }
-
-    // If split-error evaluation found candidates, return the best
-    if best_candidate.is_some() || fallback_candidate.is_some() {
-        return Ok(best_candidate.or(fallback_candidate));
-    }
-
-    // v0.1.136: If split-error evaluation was properly attempted (both subsets
-    // had enough samples) but found NOTHING, don't fall back to all-samples.
-    // The fact that neither subset produced candidates with positive net improvement
-    // means there's no reliable weight - any weight that helps one subset hurts
-    // the other by at least as much. All-samples would produce unreliable
-    // small-improvement predictions that fail in production.
-    if split_error_attempted {
-        return Ok(None);
-    }
-
-    // Fall back to original ALL-samples evaluation ONLY for cases where errors
-    // aren't clearly split (e.g., all positive or all negative errors, or
-    // one subset has too few samples)
-
-    for &orientation in spec.orientations {
-        for &scale in spec.scales {
-            let incoming_weight = orientation * scale;
-            let (
-                sum_activation_sq,
-                sum_error_activation,
-                gpu_baseline_sq,
-                _gpu_improved_count, // Ignored - calculated after weight validation
-                gpu_succeeded,
-            ) = match gpu.evaluate_activation(samples, activation_type, orientation, scale) {
-                Ok(result) => (result.0, result.1, result.2, result.3, true),
-                Err(_) => {
-                    // Fall back to CPU if GPU fails
-                    let mut sum_activation_sq = 0.0;
-                    let mut sum_error_activation = 0.0;
-                    for sample in samples {
-                        let pre_activation = incoming_weight * sample.activation;
-                        let output = (spec.activation)(pre_activation);
-                        if output.is_finite() {
-                            sum_activation_sq += output * output;
-                            sum_error_activation += output * sample.avg_error;
-                        }
-                    }
-                    (
-                        sum_activation_sq,
-                        sum_error_activation,
-                        total_baseline_error_sq,
-                        0,
-                        false,
-                    )
-                }
-            };
-
-            // Use GPU baseline if GPU succeeded, otherwise use CPU baseline
-            let baseline_sq = if gpu_succeeded {
-                gpu_baseline_sq
-            } else {
-                total_baseline_error_sq
-            };
-
-            let total_count = samples.len() as u32;
-            if total_count == 0 {
-                continue;
-            }
-
-            // For non-linear targets (HARD_TANH, ReLU, etc.), search for best outgoing_weight
-            // since linear optimal may be wrong due to activation saturation/clipping.
-            // Must verify samples have target_value/target_activation data before simulation.
-            let target_activation_fn = get_target_simulation_fn(samples, target_squash);
-            let (
-                outgoing_weight,
-                optimal_bias,
-                neuron_error_improvement, // Issue #128: Renamed - this is neuron-level, not creature-level
-                final_improved_count,
-            ) = if spec.name == "IDENTITY" {
-                let (outgoing_weight, optimal_bias) =
-                    match calculate_optimal_identity_outgoing_and_bias(samples, incoming_weight) {
-                        Some((w, b)) => (w, b),
-                        None => continue,
-                    };
-
-                let (improvement, improved_count, _) = compute_activation_improvement_and_count(
-                    samples,
-                    incoming_weight,
-                    outgoing_weight,
-                    optimal_bias,
-                    spec.activation,
-                    baseline_sq,
-                    target_activation_fn,
-                );
-
-                (outgoing_weight, optimal_bias, improvement, improved_count)
-            } else if target_activation_fn.is_some() {
-                // Use shared weight calculation with validation.
-                // For non-linear targets, we'll search over scaled versions of this base weight.
-                let base_weight = match calculate_optimal_outgoing_weight(
-                    sum_error_activation,
-                    sum_activation_sq,
-                    incoming_weight,
-                ) {
-                    Some(w) => w,
-                    None => continue, // Skip if weight is invalid or ratio too small
-                };
-
-                // Weight candidates: base weight and scaled versions
-                // All candidates are already within MAX_OUTGOING_WEIGHT since base_weight is clamped
-                let weight_candidates: [f32; 9] = [
-                    base_weight * 0.1,
-                    base_weight * 0.25,
-                    base_weight * 0.5,
-                    base_weight * 0.75,
-                    base_weight,
-                    base_weight * 1.5,
-                    base_weight * 2.0,
-                    -base_weight * 0.5,
-                    -base_weight,
-                ];
-
-                let mut best_weight = base_weight;
-                let mut best_bias = 0.0f32;
-                let mut best_improvement = f32::NEG_INFINITY;
-                let mut best_improved_count = 0u32;
-
-                for &weight in &weight_candidates {
-                    // Clamp scaled weights to ensure they stay within bounds
-                    let clamped_weight = weight.clamp(-MAX_OUTGOING_WEIGHT, MAX_OUTGOING_WEIGHT);
-                    if clamped_weight.abs() <= EPSILON {
-                        continue;
-                    }
-
-                    let bias = calculate_optimal_bias(
-                        samples,
-                        incoming_weight,
-                        clamped_weight,
-                        spec.activation,
-                        spec.name,
-                        None,
-                        target_squash,
-                    );
-
-                    // v0.1.122: When target simulation is available, DON'T recompute weight
-                    // using VALUE domain optimization. The candidate weights are already
-                    // scaled versions of linear optimal. Recomputing in VALUE domain can
-                    // overshoot near saturation, leading to worse ACTIVATION domain results.
-                    // Instead, let the ACTIVATION domain evaluation pick the best candidate.
-                    //
-                    // We DO recompute to account for bias changing the activation pattern,
-                    // but only when we don't have target simulation (linear approximation).
-
-                    // Single pass for improvement and count with target simulation
-                    let (improvement, improved, _) = compute_activation_improvement_and_count(
-                        samples,
-                        incoming_weight,
-                        clamped_weight, // Use candidate weight directly, not recomputed
-                        bias,
-                        spec.activation,
-                        baseline_sq,
-                        target_activation_fn,
-                    );
-
-                    if improvement > best_improvement {
-                        best_improvement = improvement;
-                        best_weight = clamped_weight; // Use candidate weight directly
-                        best_bias = bias;
-                        best_improved_count = improved;
-                    }
-                }
-
-                (
-                    best_weight,
-                    best_bias,
-                    best_improvement,
-                    best_improved_count,
-                )
-            } else {
-                // For linear targets or when target data unavailable, use the base weight.
-                let base_weight = match calculate_optimal_outgoing_weight(
-                    sum_error_activation,
-                    sum_activation_sq,
-                    incoming_weight,
-                ) {
-                    Some(w) => w,
-                    None => continue, // Skip if weight is invalid or ratio too small
-                };
-
-                // For linear targets or when target data unavailable, use the base weight
-                let optimal_bias = calculate_optimal_bias(
-                    samples,
-                    incoming_weight,
-                    base_weight,
-                    spec.activation,
-                    spec.name,
-                    None,
-                    target_squash,
-                );
-
-                // CRITICAL FIX: Recompute optimal weight WITH the bias included.
-                // The base_weight was computed WITHOUT bias, so recompute
-                // using the actual activation pattern with bias.
-                let mut sum_activation_sq_with_bias = 0.0f32;
-                let mut sum_error_activation_with_bias = 0.0f32;
-                for sample in samples.iter() {
-                    let pre_activation = incoming_weight * sample.activation + optimal_bias;
-                    let output = (spec.activation)(pre_activation);
-                    if output.is_finite() {
-                        sum_activation_sq_with_bias += output * output;
-                        sum_error_activation_with_bias += output * sample.avg_error;
-                    }
-                }
-                // Use shared function for bias-adjusted weight calculation
-                let outgoing_weight = calculate_optimal_outgoing_weight(
-                    sum_error_activation_with_bias,
-                    sum_activation_sq_with_bias,
-                    incoming_weight,
-                )
-                .unwrap_or(base_weight);
-
-                let (improvement, improved_count, _) = compute_activation_improvement_and_count(
-                    samples,
-                    incoming_weight,
-                    outgoing_weight,
-                    optimal_bias,
-                    spec.activation,
-                    baseline_sq,
-                    None, // Linear approximation
-                );
-
-                (outgoing_weight, optimal_bias, improvement, improved_count)
-            };
-
-            // Skip invalid weights
-            if outgoing_weight.abs() <= EPSILON {
-                continue;
-            }
-
-            // =================================================================
-            // FUNDAMENTAL VALIDITY FILTERS
-            // These filters apply to ALL candidates (both fallback and best).
-            // They must be checked BEFORE setting fallback_candidate to prevent
-            // invalid candidates from being returned through the fallback path.
-            // =================================================================
-
-            // Issue #123: Check for saturation - reject if neuron output is nearly constant.
-            // Large bias values (e.g., 5.0 for SOFTSIGN) can cause the neuron to saturate,
-            // producing nearly identical output regardless of input. Such neurons cannot
-            // reduce error correlation and lead to massive prediction failures.
-            if !has_sufficient_output_variance(
-                samples,
-                incoming_weight,
-                optimal_bias,
-                spec.activation,
-            ) {
-                continue;
-            }
-
-            // Require minimum ABSOLUTE error reduction, not just percentage.
-            // A 1% improvement on baseline_sq=0.0001 is only 0.000001 absolute reduction,
-            // which won't meaningfully affect the creature's total error.
-            // Minimum absolute improvement = 0.001 (0.1% of typical baseline ~1.0)
-            let absolute_improvement = neuron_error_improvement * baseline_sq;
-            if absolute_improvement < 0.001 {
-                continue;
-            }
-
-            // IDENTITY with bias ≈ 0 is equivalent to a direct synapse (source × incoming × outgoing)
-            // Filter out these redundant candidates - use synapse analysis for direct connections
-            if spec.name == "IDENTITY" && optimal_bias.abs() < 0.01 {
-                continue;
-            }
-
-            // =================================================================
-            // FALLBACK CANDIDATE (best seen so far with any positive improvement)
-            // =================================================================
-            // Track the best candidate that didn't pass the main threshold.
-            // TypeScript will decide if the improvement is worth the cost of growth.
-            // We don't apply arbitrary minimum thresholds here - any positive
-            // improvement is returned and TypeScript handles candidate selection.
-            if neuron_error_improvement > fallback_score && neuron_error_improvement > 0.0 {
-                fallback_score = neuron_error_improvement;
-
-                let target_stats = NeuronStats::from_samples(samples).map(|s| s.to_json());
-                // Issue #128: Use creature-level metrics (impact discounting applied later)
-                fallback_candidate = Some(CandidateNeuronJson {
-                    source_neuron_uuid: source_uuid.to_string(),
-                    target_neuron_uuid: target_uuid.to_string(),
-                    source_neuron_index: None, // Set during impact discounting
-                    target_neuron_index: None, // Set during impact discounting
-                    incoming_weight,
-                    outgoing_weight,
-                    squash: spec.name.to_string(),
-                    bias: optimal_bias,
-                    comment: None,
-                    target_neuron_impact: 1.0,
-                    expected_creature_error_reduction: neuron_error_improvement,
-                    expected_creature_score_gain: neuron_error_improvement,
-                    improved_count: final_improved_count,
-                    total_count,
-                    target_neuron_stats: target_stats,
-                });
-            }
-
-            // =================================================================
-            // THRESHOLD CHECK (only for best_candidate, not fallback)
-            // =================================================================
-            // Use the STRICTER threshold (max) to ensure meaningful improvements
-            // If spec requires 5% and caller passes 1%, we require 5%
-            // If spec requires 0% and caller passes 1%, we require 1%
-            let improvement_cutoff = threshold.max(spec.min_improvement);
-
-            if neuron_error_improvement <= improvement_cutoff
-                || final_improved_count < MIN_NEURON_SAMPLE_COUNT as u32
-            {
-                continue;
-            }
-
-            // Current iteration passed threshold - create best_candidate with current iteration's values
-            if neuron_error_improvement > best_score {
-                best_score = neuron_error_improvement;
-
-                let target_stats = NeuronStats::from_samples(samples).map(|s| s.to_json());
-                // Issue #128: Use creature-level metrics (impact discounting applied later)
-                best_candidate = Some(CandidateNeuronJson {
-                    source_neuron_uuid: source_uuid.to_string(),
-                    target_neuron_uuid: target_uuid.to_string(),
-                    source_neuron_index: None, // Set during impact discounting
-                    target_neuron_index: None, // Set during impact discounting
-                    incoming_weight,
-                    outgoing_weight,
-                    squash: spec.name.to_string(),
-                    bias: optimal_bias,
-                    comment: None,
-                    target_neuron_impact: 1.0,
-                    expected_creature_error_reduction: neuron_error_improvement,
-                    expected_creature_score_gain: neuron_error_improvement,
-                    improved_count: final_improved_count,
-                    total_count,
-                    target_neuron_stats: target_stats,
-                });
-            }
-        }
-    }
-
-    Ok(best_candidate.or(fallback_candidate))
-}
-
-// Threshold-crossing (STEP/BIPOLAR) sample building was removed as dead code.
-// For discrete targets, add-synapse analysis is the intended mechanism.
-
-// Threshold-crossing (STEP/BIPOLAR) add-neuron evaluation was removed as dead code.
-// For discrete targets, add-synapse analysis is the intended mechanism.
-
-// FocusTargetFilterResult, filter_focus_targets_for_neuron_analysis
-// moved to crate::analysis::diagnostics module (Issue #271)
-
-pub(crate) fn analyze_neurons_with_cache(
-    input: &AnalyzeNeuronsInput,
-    cache: Arc<RecordCache>,
-) -> Result<AnalyzeNeuronsResult> {
-    // v0.1.134: Return ALL positive improvements.
-    // NEAT-AI applies the cost-of-growth gate during evaluation.
-    let threshold = 0.0;
-    let ordered_neurons = build_ordered_neurons(&input.creature);
-
-    // Build a lookup map for neuron squash functions to identify discrete targets
-    let neuron_squash_map: HashMap<String, String> = input
-        .creature
-        .neurons
-        .iter()
-        .map(|n| (n.uuid.clone(), n.squash.clone()))
-        .collect();
-
-    // Build a comprehensive lookup map for ALL neuron UUIDs to their types.
-    // This includes: input neurons (from creature.input count) and all neurons from
-    // creature.neurons (hidden, output, constant). If a UUID is not in this map,
-    // it's an invalid UUID (bug in the caller).
-    //
-    // Only output neurons should be targets for add-neuron candidates because:
-    // - Output neuron errors directly affect creature score
-    // - Hidden neuron errors are backpropagated approximations that don't correlate
-    //   reliably with actual output error reduction
-    // - Input neurons are observation sources, not computation nodes
-    let mut neuron_type_map: HashMap<String, String> = HashMap::new();
-
-    // Add input neurons (they're not in creature.neurons, only represented by creature.input count)
-    for input_index in 0..input.creature.input {
-        neuron_type_map.insert(format!("input-{input_index}"), "input".to_string());
-    }
-
-    // Add all neurons from creature.neurons (hidden, output, constant)
-    for neuron in &input.creature.neurons {
-        neuron_type_map.insert(neuron.uuid.clone(), neuron.neuron_type.clone());
-    }
-
-    // Log creature configuration for debugging data issues
-    if verbose_enabled() {
-        let non_input_count = input.creature.neurons.len();
-        let total_neurons = input.creature.input + non_input_count;
-        eprintln!(
-            "[NEAT-AI-Discovery][verbose] Neuron analysis creature config: {} input neurons (input-0 to input-{}), {} non-input neurons, {} total ordered neurons",
-            input.creature.input,
-            input.creature.input.saturating_sub(1),
-            non_input_count,
-            total_neurons
-        );
-
-        // Verify input neurons exist in parquet by checking a sample
-        if input.creature.input > 0 {
-            match cache.get("input-0") {
-                Ok(records) => {
-                    eprintln!(
-                        "[NEAT-AI-Discovery][verbose] Parquet data check: input-0 has {} records",
-                        records.len()
-                    );
-                    if !records.is_empty() {
-                        let first = &records[0];
-                        let last = &records[records.len() - 1];
-                        eprintln!(
-                            "[NEAT-AI-Discovery][verbose] Parquet data check: input-0 obs_index range [{}, {}], first activation={:.4}",
-                            first.obs_index,
-                            last.obs_index,
-                            first.activation
-                        );
-                    }
-                }
-                Err(err) => {
-                    eprintln!(
-                        "[NEAT-AI-Discovery][verbose] Parquet data check FAILED: input-0 error: {err}"
-                    );
-                }
-            }
-
-            // Also check a middle input neuron
-            let mid_input = input.creature.input / 2;
-            let mid_uuid = format!("input-{mid_input}");
-            match cache.get(&mid_uuid) {
-                Ok(records) => {
-                    eprintln!(
-                        "[NEAT-AI-Discovery][verbose] Parquet data check: {mid_uuid} has {} records",
-                        records.len()
-                    );
-                }
-                Err(err) => {
-                    eprintln!(
-                        "[NEAT-AI-Discovery][verbose] Parquet data check FAILED: {mid_uuid} error: {err}"
-                    );
-                }
-            }
-        }
-    }
-
-    let order_map: HashMap<String, usize> = ordered_neurons
-        .iter()
-        .map(|neuron| (neuron.uuid.clone(), neuron.index))
-        .collect();
-
-    let unique_focus = require_unique_focus(&input.focus_neurons, "analyse_neurons")?;
-
-    let helpful_map = Arc::new(Mutex::new(HashMap::<
-        (String, String, String, i8, i8),
-        CandidateNeuronJson,
-    >::new()));
-
-    let diagnostics = Arc::new(Mutex::new(NeuronDiagnostics::new(&unique_focus)));
-
-    // GPU timing collector (Issue #195)
-    // Only collects timing data when NEAT_AI_DISCOVERY_GPU_TIMING=1 is set
-    let timing_collector = Arc::new(super::shared::TimingCollector::new(
-        super::utils::gpu_timing_enabled(),
-    ));
-
-    let deadline = build_deadline(input.analysis_deadline_ms);
-    // GPU is always required - TypeScript layer calls check_gpu_available() and skips
-    // discovery entirely on machines without GPU. Reaching here without GPU is a bug.
-    assert!(
-        GpuAnalyzer::gpu_is_available(),
-        "Discovery logic called without GPU - check_gpu_available should have prevented this"
-    );
-    let gpu_used = true;
-    let analysis_timed_out = Arc::new(Mutex::new(false));
-
-    // Filter focus neurons to valid add-neuron targets.
-    //
-    // Rationale: Add-neuron candidates predict error reduction at the target neuron.
-    // For OUTPUT neurons, this directly corresponds to creature score improvement.
-    // For HIDDEN neurons, the backpropagated error is an approximation that doesn't
-    // reliably translate to actual output error reduction - we've observed 100%
-    // failure rates when targeting hidden neurons.
-    // For INPUT neurons, they are observation sources, not computation nodes - they
-    // have no activation function or error to reduce.
-    //
-    // Non-output neurons are skipped here but still tracked in diagnostics so the
-    // caller knows they were received but filtered out (with the correct reason).
-    let original_focus_count = unique_focus.len();
-    // Optional production experiment (29-Dec-2025): allow callers to force output-only
-    // focus targets for add-neuron analysis.
-    //
-    // This is disabled by default for backwards compatibility (existing regression tests
-    // and older pipelines expect hidden targets to be analysed with impact discounting).
-    //
-    // Enable by setting `NEAT_AI_DISCOVERY_NEURON_TARGETS_OUTPUT_ONLY=1`.
-    let output_only_targets = std::env::var("NEAT_AI_DISCOVERY_NEURON_TARGETS_OUTPUT_ONLY").is_ok();
-    let FocusTargetFilterResult {
-        mut focus_order,
-        skipped_hidden,
-        skipped_input,
-        skipped_constant,
-        threshold_targets,
-    } = filter_focus_targets_for_neuron_analysis(
-        &unique_focus,
-        &neuron_type_map,
-        &neuron_squash_map,
-        output_only_targets,
-    );
-
-    // Log when non-output neurons are filtered out
-    let total_skipped = skipped_hidden.len() + skipped_input.len() + skipped_constant.len();
-    if total_skipped > 0 {
-        // Build a summary of skipped neuron types
-        let mut skipped_parts: Vec<String> = Vec::new();
-        if !skipped_input.is_empty() {
-            skipped_parts.push(format!(
-                "Input: {:?}",
-                skipped_input.iter().take(5).collect::<Vec<_>>()
-            ));
-        }
-        if !skipped_hidden.is_empty() {
-            skipped_parts.push(format!(
-                "Hidden: {:?}",
-                skipped_hidden.iter().take(5).collect::<Vec<_>>()
-            ));
-        }
-        if !skipped_constant.is_empty() {
-            skipped_parts.push(format!(
-                "Constant: {:?}",
-                skipped_constant.iter().take(5).collect::<Vec<_>>()
-            ));
-        }
-        eprintln!(
-            "[NEAT-AI-Discovery] Filtered {} neuron(s) from add-neuron analysis. {}. Remaining valid targets: {}",
-            total_skipped,
-            skipped_parts.join(". "),
-            focus_order.len()
-        );
-    }
-
-    // If no output neurons remain after filtering, return early with empty results
-    if focus_order.is_empty() {
-        eprintln!(
-            "[NEAT-AI-Discovery] No output neurons in focus list ({original_focus_count} non-output neurons filtered out). \
-            Add-neuron candidates can only target output neurons."
-        );
-        // Build no_candidate_reasons with correct reason for each neuron type
-        let mut no_candidate_reasons: Vec<NeuronNoCandidateSummary> = Vec::new();
-        for uuid in &skipped_input {
-            no_candidate_reasons.push(NeuronNoCandidateSummary {
-                target_uuid: uuid.clone(),
-                reason: NeuronNoCandidateReason::InputNeuronFiltered,
-                evaluated_sources: 0,
-                sources_with_samples: 0,
-                target_record_count: 0,
-                detail: None,
-            });
-        }
-        for uuid in &skipped_hidden {
-            no_candidate_reasons.push(NeuronNoCandidateSummary {
-                target_uuid: uuid.clone(),
-                reason: NeuronNoCandidateReason::HiddenNeuronFiltered,
-                evaluated_sources: 0,
-                sources_with_samples: 0,
-                target_record_count: 0,
-                detail: None,
-            });
-        }
-        for uuid in &skipped_constant {
-            no_candidate_reasons.push(NeuronNoCandidateSummary {
-                target_uuid: uuid.clone(),
-                reason: NeuronNoCandidateReason::ConstantNeuronFiltered,
-                evaluated_sources: 0,
-                sources_with_samples: 0,
-                target_record_count: 0,
-                detail: None,
-            });
-        }
-        return Ok(AnalyzeNeuronsResult {
-            helpful_neurons: Vec::new(),
-            gpu_used: true,
-            no_candidate_reasons,
-            metadata: super::shared::NeuronAnalysisMetadata {
-                candidates_found: 0,
-                candidates_returned: 0,
-                timed_out: false,
-                completed_focus_neurons: 0,
-                total_focus_neurons: original_focus_count,
-                timing: None,
-                gpu_info: GpuAnalyzer::get_adapter_info(),
-            },
-        });
-    }
-
-    // Mark skipped neurons in diagnostics so they appear with the correct reason
-    // instead of misleading reasons like NoEligibleSources.
-    // This is the normal flow case where some output neurons exist.
-    for input_uuid in &skipped_input {
-        diagnostics
-            .lock()
-            .expect("Mutex poisoned: diagnostics")
-            .mark_input_filtered(input_uuid);
-    }
-    for hidden_uuid in &skipped_hidden {
-        diagnostics
-            .lock()
-            .expect("Mutex poisoned: diagnostics")
-            .mark_hidden_filtered(hidden_uuid);
-    }
-    for constant_uuid in &skipped_constant {
-        diagnostics
-            .lock()
-            .expect("Mutex poisoned: diagnostics")
-            .mark_constant_filtered(constant_uuid);
-    }
-
-    // Log threshold-crossing neurons for visibility
-    if verbose_enabled() && !threshold_targets.is_empty() {
-        eprintln!(
-            "[NEAT-AI-Discovery][verbose] Using threshold-crossing model for {} STEP/BIPOLAR neurons: {:?}",
-            threshold_targets.len(),
-            threshold_targets.iter().take(5).collect::<Vec<_>>()
-        );
-    }
-
-    shuffle_slice(&mut focus_order, input.random_seed, "neuron:focus_order");
-
-    // Log analysis start with timeout duration and randomised order
-    log_analysis_start(
-        "neuron",
-        input.analysis_deadline_ms,
-        focus_order.len(),
-        &focus_order,
-    );
-
-    // Track completed focus neurons for timeout logging
-    let total_focus_count = focus_order.len();
-    let completed_count = Arc::new(std::sync::atomic::AtomicUsize::new(0));
-
-    let focus_order_arc = Arc::new(focus_order);
-    let ordered_neurons_arc = Arc::new(ordered_neurons);
-    let order_map_arc = Arc::new(order_map);
-    let neuron_squash_map_arc = Arc::new(neuron_squash_map);
-
-    // Issue #182: Build a set of "used" input neurons (those with at least one outgoing synapse)
-    // for the focus_unused_observations feature. Unused inputs will be prioritised in source ordering.
-    let used_inputs: HashSet<String> = input
-        .creature
-        .synapses
-        .iter()
-        .filter(|s| parse_input_index(&s.from_uuid).is_some())
-        .map(|s| s.from_uuid.clone())
-        .collect();
-    let used_inputs_arc = Arc::new(used_inputs);
-
-    // Create a shared GPU work queue ONCE before the parallel loop.
-    // This eliminates the overhead of creating multiple GPU devices (one per thread).
-    // All GPU operations are processed by a single dedicated thread, improving utilisation.
-    // CRITICAL: The GpuAnalyzer is created INSIDE the GPU thread to avoid wgpu deadlocks.
-    let gpu_queue = Arc::new(GpuWorkQueue::new()?);
-
-    // Process each focus neuron in parallel. Deadline checks happen at the start of
-    // each focus target so that once analysis for a neuron begins, we prefer to
-    // complete its upstream evaluation rather than abandoning it mid-stream. This
-    // gives us “vertical” timeout behaviour where some neurons complete fully even
-    // if later targets are skipped when the deadline is reached.
-    focus_order_arc
-        .par_iter()
-        .try_for_each(|target_uuid| -> Result<()> {
-            if *analysis_timed_out.lock().expect("Mutex poisoned") || deadline_passed(&deadline) {
-                *analysis_timed_out.lock().expect("Mutex poisoned") = true;
-                return Ok(());
-            }
-
-            crate::watchdog::beat(format!(
-                "neuron analysis → processing target {target_uuid}"
-            ));
-
-            // Use the shared GPU work queue instead of creating a new GpuAnalyzer per thread.
-            // This eliminates device creation overhead and improves GPU utilisation.
-            let gpu = &*gpu_queue;
-
-            let target_records_arc = match cache.get(target_uuid.as_str()) {
-                Ok(records) => records,
-                Err(err) => {
-                    if cfg!(debug_assertions) {
-                        eprintln!("Failed to load target neuron records for {target_uuid}: {err}");
-                    }
-                    return Ok(());
-                }
-            };
-            if target_records_arc.is_empty() {
-                diagnostics
-                    .lock()
-                    .expect("Mutex poisoned: diagnostics")
-                    .set_target_record_count(target_uuid, 0);
-                return Ok(());
-            }
-            let target_records = target_records_arc.as_ref();
-            diagnostics
-                .lock()
-                .expect("Mutex poisoned: diagnostics")
-                .set_target_record_count(target_uuid, target_records.len());
-
-            // Log target neuron obs_index range for debugging sample matching
-            if verbose_enabled() && !target_records.is_empty() {
-                let first_obs = target_records.first().map(|r| r.obs_index).unwrap_or(0);
-                let last_obs = target_records.last().map(|r| r.obs_index).unwrap_or(0);
-                let has_errors = target_records.iter().any(|r| !r.errors.is_empty());
-                eprintln!(
-                    "[NEAT-AI-Discovery][verbose] Target {} has {} records, obs_index range [{}, {}], has_errors={}",
-                    target_uuid,
-                    target_records.len(),
-                    first_obs,
-                    last_obs,
-                    has_errors
-                );
-            }
-
-            // STEP/BIPOLAR are discrete targets. Add-neuron discovery for these targets was
-            // removed as dead code; add-synapse is the intended mechanism.
-            let is_threshold_target = neuron_squash_map_arc
-                .get(target_uuid)
-                .map(|squash| is_threshold_activation(squash))
-                .unwrap_or(false);
-
-            let target_index = match order_map_arc.get(target_uuid.as_str()) {
-                Some(index) => *index,
-                None => return Ok(()),
-            };
-
-            let mut eligible_sources: Vec<&OrderedNeuron> = ordered_neurons_arc
-                .iter()
-                .filter(|neuron| neuron.index < target_index)
-                .collect();
-            // Source ordering:
-            // - Candidates must remain forward-only (index < target_index).
-            // - Under timeouts we want coverage over time, so we fully shuffle ALL
-            //   eligible sources (inputs + hidden + constants).
-            // - If NEAT_AI_DISCOVERY_FOCUS_UNUSED_OBSERVATIONS=1, unused inputs are
-            //   prioritised to the front (Issue #182).
-            //
-            // If `random_seed` is provided, the shuffle is deterministic for a given
-            // target UUID; otherwise it's non-deterministic.
-            let context = format!("neuron:eligible_sources:{target_uuid}");
-            order_eligible_sources(
-                &mut eligible_sources,
-                input.random_seed,
-                &context,
-                input.creature.input,
-                Some(&*used_inputs_arc),
-            );
-
-            // Track total eligible sources for diagnostics
-            let total_eligible = eligible_sources.len() as u32;
-            diagnostics
-                .lock()
-                .expect("Mutex poisoned: diagnostics")
-                .set_total_eligible_sources(target_uuid, total_eligible);
-
-            // Log focus neuron details for debugging
-            if verbose_enabled() && total_eligible == 0 {
-                eprintln!(
-                    "[NEAT-AI-Discovery][verbose] Target {} (index {}) has 0 eligible upstream sources. creature.input={}, so input neurons span indices 0-{}. This indicates target_index <= 0 or a creature configuration mismatch.",
-                    target_uuid,
-                    target_index,
-                    ordered_neurons_arc.len().saturating_sub(input.creature.neurons.len()),
-                    ordered_neurons_arc.len().saturating_sub(input.creature.neurons.len()).saturating_sub(1)
-                );
-            }
-
-            // Phase 1: Pre-filter sources and collect their records (with deadline checks)
-            // This mirrors the synapse analysis approach for better parallelism
-            let mut sources_to_process: Vec<(&OrderedNeuron, Arc<Vec<DiscoverRecord>>)> =
-                Vec::with_capacity(eligible_sources.len());
-            let mut empty_record_sources: Vec<String> = Vec::new();
-            let mut load_failure_count = 0u32;
-
-            for source in &eligible_sources {
-                // Check deadline during pre-filtering
-                if deadline_passed(&deadline) {
-                    *analysis_timed_out.lock().expect("Mutex poisoned") = true;
-                    break;
-                }
-                let source_uuid = source.uuid.as_str();
-                match cache.get(source_uuid) {
-                    Ok(records) => {
-                        if !records.is_empty() {
-                            sources_to_process.push((source, records));
-                        } else {
-                            empty_record_sources.push(source_uuid.to_string());
-                        }
-                    }
-                    Err(err) => {
-                        load_failure_count += 1;
-                        if verbose_enabled() {
-                            eprintln!(
-                                "[NEAT-AI-Discovery][verbose] Failed to load source neuron records for {source_uuid} (target {target_uuid}): {err}"
-                            );
-                        }
-                    }
-                }
-            }
-
-            // Record load failures in diagnostics
-            if load_failure_count > 0 {
-                let mut diag = diagnostics.lock().expect("Mutex poisoned: diagnostics");
-                for _ in 0..load_failure_count {
-                    diag.record_load_failure(target_uuid);
-                }
-            }
-
-            // Log summary of source loading results for debugging
-            let sources_checked = sources_to_process.len() + empty_record_sources.len() + load_failure_count as usize;
-            let timed_out_during_loading = *analysis_timed_out.lock().expect("Mutex poisoned");
-            if verbose_enabled() && (sources_to_process.is_empty() || load_failure_count > 0 || !empty_record_sources.is_empty() || timed_out_during_loading) {
-                let sources_with_records = sources_to_process.len();
-                let empty_count = empty_record_sources.len();
-                if timed_out_during_loading && sources_checked == 0 {
-                    eprintln!(
-                        "[NEAT-AI-Discovery][verbose] Target {target_uuid} source loading: TIMEOUT before any of {total_eligible} eligible sources could be checked"
-                    );
-                } else if timed_out_during_loading {
-                    eprintln!(
-                        "[NEAT-AI-Discovery][verbose] Target {target_uuid} source loading: TIMEOUT after checking {sources_checked}/{total_eligible} eligible sources ({sources_with_records} with records, {empty_count} empty, {load_failure_count} failures)"
-                    );
-                } else {
-                    eprintln!(
-                        "[NEAT-AI-Discovery][verbose] Target {target_uuid} source loading: {total_eligible} eligible -> {sources_with_records} with records, {empty_count} empty records, {load_failure_count} load failures"
-                    );
-                }
-            }
-
-            // Batch diagnostics for empty record sources
-            if !empty_record_sources.is_empty() {
-                let mut diag = diagnostics.lock().expect("Mutex poisoned: diagnostics");
-                for source_uuid in &empty_record_sources {
-                    diag.record_candidate_attempt(target_uuid, false);
-                    diag.record_no_samples(target_uuid, source_uuid);
-                }
-            }
-
-            // Check if timed out during pre-filtering
-            if *analysis_timed_out.lock().expect("Mutex poisoned") {
-                return Ok(());
-            }
-
-            // Phase 2: Build samples in parallel using CPU (much faster than sequential GPU calls)
-            // OPTIMIZATION: Pre-build target map ONCE, reuse for all sources.
-            // This avoids rebuilding the HashMap for each of ~1000+ source neurons.
-            let target_map = TargetMap::from_records(target_records);
-            let target_map_ref = &target_map;
-
-            // Even if target_map is empty, we continue to record diagnostics
-            // about what sources were evaluated.
-            struct NeuronWorkResult {
-                source_uuid: String,
-                samples: Vec<HelpfulSample>,
-            }
-
-            // Issue #221: Sample Locality Optimisation
-            // Group sources by obs_index overlap to reduce redundant sample building.
-            // Sources with ≥80% obs_index overlap share sample building in a single pass.
-            let work_results: Vec<NeuronWorkResult> = {
-                let _timing = TimingScope::sample_building(&timing_collector);
-
-                // Group sources by sample locality
-                let locality_groups = group_sources_by_locality(&sources_to_process);
-
-                // Log locality grouping stats if verbose
-                if verbose_enabled() && sources_to_process.len() >= MIN_GROUP_SIZE_FOR_LOCALITY {
-                    let group_sizes: Vec<usize> = locality_groups.iter().map(|g| g.sources.len()).collect();
-                    let max_group = group_sizes.iter().max().copied().unwrap_or(0);
-                    let avg_group = if !group_sizes.is_empty() {
-                        group_sizes.iter().sum::<usize>() as f32 / group_sizes.len() as f32
-                    } else {
-                        0.0
-                    };
-                    eprintln!(
-                        "[NEAT-AI-Discovery][verbose] Neuron analysis target {}: {} sources grouped into {} locality groups (max={}, avg={:.1})",
-                        target_uuid,
-                        sources_to_process.len(),
-                        locality_groups.len(),
-                        max_group,
-                        avg_group
-                    );
-                }
-
-                // Build samples for each group (groups with multiple sources use batched building)
-                locality_groups
-                    .par_iter()
-                    .flat_map(|group| {
-                        let group_results = build_samples_for_locality_group(group, target_map_ref);
-                        group_results
-                            .into_iter()
-                            .map(|(source_uuid, samples, _record_count)| NeuronWorkResult {
-                                source_uuid,
-                                samples,
-                            })
-                            .collect::<Vec<_>>()
-                    })
-                    .collect()
-            };
-
-            // Phase 3: Batch diagnostics updates for sample building results
-            {
-                let mut diag = diagnostics.lock().expect("Mutex poisoned: diagnostics");
-                for result in &work_results {
-                    diag.record_candidate_attempt(target_uuid, !result.samples.is_empty());
-                    if result.samples.is_empty() {
-                        diag.record_no_samples(target_uuid, &result.source_uuid);
-                    }
-                }
-            }
-
-            // Phase 4: Process evaluations - GPU work is done here.
-            // Filter to only sources with samples, then evaluate.
-            //
-            // NOTE: We intentionally skip STEP/BIPOLAR targets here; synapse analysis
-            // is the supported discovery mechanism for discrete targets.
-            //
-            // Important: We still count the target as "completed" for progress reporting.
-            if !is_threshold_target {
-                // Standard continuous activation path
-                for result in work_results {
-                    // Check deadline before each evaluation batch
-                    if deadline_passed(&deadline) {
-                        *analysis_timed_out.lock().expect("Mutex poisoned") = true;
-                        break;
-                    }
-
-                    if result.samples.is_empty() {
-                        continue;
-                    }
-
-                    // Get target_squash for accurate HARD_TANH modelling
-                    let target_squash = neuron_squash_map_arc.get(target_uuid).map(|s| s.as_str());
-
-                    // Issue #130 (v0.2.2): Compute source variance discount.
-                    // If source activation has low variance, predictions are unreliable.
-                    let source_variance_discount = compute_source_variance_discount(&result.samples);
-                    if source_variance_discount <= EPSILON {
-                        // Source is constant - skip evaluation entirely
-                        continue;
-                    }
-
-                    // ReLU evaluation: split by TARGET neuron's error sign.
-                    //
-                    // ReLU can only push output in ONE direction (based on outgoing weight sign),
-                    // so we evaluate two candidates separately:
-                    // - Positive-error ReLU: optimised for samples where output should be HIGHER
-                    // - Negative-error ReLU: optimised for samples where output should be LOWER
-                    //
-                    // Each candidate's weight is computed from its error subset, then NET
-                    // improvement is calculated across ALL samples. This is the correct
-                    // approach for directional activation functions like ReLU.
-                    //
-                    // NOTE: We don't use "averaging over all samples" because when errors are
-                    // split ~50/50, the average cancels out and no candidate is found.
-                    let split_result = {
-                        let _timing = TimingScope::shader(&timing_collector, "relu");
-                        evaluate_relu_candidates_split(
-                            gpu,
-                            &result.source_uuid,
-                            target_uuid,
-                            &result.samples,
-                            threshold,
-                            target_squash,
-                        )?
-                    };
-
-                    if let Some(mut candidate) = split_result.positive_error_candidate {
-                        // Issue #130: Apply source variance discount
-                        candidate.expected_creature_error_reduction *= source_variance_discount;
-                        candidate.expected_creature_score_gain *= source_variance_discount;
-
-                        if verbose_enabled() {
-                            eprintln!(
-                                "[NEAT-AI-Discovery][verbose] ReLU (push UP) {} -> {}: {:.2}% improvement (variance discount: {:.2})",
-                                result.source_uuid,
-                                target_uuid,
-                                candidate.expected_creature_score_gain * 100.0,
-                                source_variance_discount
-                            );
-                        }
-                        diagnostics
-                            .lock()
-                            .expect("Mutex poisoned: diagnostics")
-                            .mark_candidate_selected(target_uuid);
-                        let mut map = helpful_map.lock().expect("Mutex poisoned: helpful_map");
-                        upsert_candidate(&mut map, candidate);
-                    }
-
-                    if let Some(mut candidate) = split_result.negative_error_candidate {
-                        // Issue #130: Apply source variance discount
-                        candidate.expected_creature_error_reduction *= source_variance_discount;
-                        candidate.expected_creature_score_gain *= source_variance_discount;
-
-                        if verbose_enabled() {
-                            eprintln!(
-                                "[NEAT-AI-Discovery][verbose] ReLU (push DOWN) {} -> {}: {:.2}% improvement (variance discount: {:.2})",
-                                result.source_uuid,
-                                target_uuid,
-                                candidate.expected_creature_score_gain * 100.0,
-                                source_variance_discount
-                            );
-                        }
-                        diagnostics
-                            .lock()
-                            .expect("Mutex poisoned: diagnostics")
-                            .mark_candidate_selected(target_uuid);
-                        let mut map = helpful_map.lock().expect("Mutex poisoned: helpful_map");
-                        upsert_candidate(&mut map, candidate);
-                    }
-
-                    for spec in ACTIVATION_SPECS.iter() {
-                        let candidate_result = {
-                            let _timing = TimingScope::shader(&timing_collector, "activation");
-                            evaluate_activation_candidate(
-                                gpu,
-                                &result.source_uuid,
-                                target_uuid,
-                                &result.samples,
-                                threshold,
-                                spec,
-                                target_squash,
-                            )?
-                        };
-                        if let Some(mut candidate) = candidate_result {
-                            // Issue #130: Apply source variance discount
-                            candidate.expected_creature_error_reduction *= source_variance_discount;
-                            candidate.expected_creature_score_gain *= source_variance_discount;
-
-                            diagnostics
-                                .lock()
-                                .expect("Mutex poisoned: diagnostics")
-                                .mark_candidate_selected(target_uuid);
-                            let mut map = helpful_map.lock().expect("Mutex poisoned: helpful_map");
-                            upsert_candidate(&mut map, candidate);
-                        }
-                    }
-                }
-            }
-
-            // Track completion of this focus neuron for timeout reporting.
-            let completed =
-                completed_count.fetch_add(1, std::sync::atomic::Ordering::Relaxed) + 1;
-            crate::watchdog::beat(format!(
-                "neuron analysis → completed {completed}/{total_focus_count} (last target {target_uuid})"
-            ));
-            Ok(())
-        })?;
-
-    let analysis_timed_out = *analysis_timed_out
-        .lock()
-        .expect("Mutex poisoned: analysis_timed_out");
-    let helpful_map = helpful_map
-        .lock()
-        .expect("Mutex poisoned: helpful_map")
-        .clone();
-    let diagnostics = diagnostics.lock().expect("Mutex poisoned: diagnostics");
-
-    // Log timeout with completion stats (always visible, not just verbose)
-    if analysis_timed_out {
-        let completed = completed_count.load(std::sync::atomic::Ordering::Relaxed);
-        log_analysis_timeout("neuron", completed, total_focus_count);
-    }
-
-    let mut helpful_results: Vec<CandidateNeuronJson> = helpful_map.into_values().collect();
-
-    // Issue #128: Apply impact-based discounting and set creature-level metrics.
-    // Output neurons have impact = 1.0 (no discount).
-    // Hidden neurons have impact in [0, 1] based on their weighted paths to outputs.
-    let impact_scores = compute_impact_scores_for_discounting(&input.creature, cache.as_ref());
-    for candidate in &mut helpful_results {
-        candidate.source_neuron_index = order_map_arc.get(&candidate.source_neuron_uuid).copied();
-        candidate.target_neuron_index = order_map_arc.get(&candidate.target_neuron_uuid).copied();
-
-        let is_hidden = neuron_type_map
-            .get(&candidate.target_neuron_uuid)
-            .map(|t| t != "output")
-            .unwrap_or(true); // Default to true if type unknown (treat as hidden)
-
-        let impact = if is_hidden {
-            if let Some(&impact) = impact_scores.get(&candidate.target_neuron_uuid) {
-                impact.clamp(0.0, 1.0)
-            } else {
-                // No impact score means disconnected from outputs - heavy discount
-                0.1
-            }
-        } else {
-            // Output neuron - full impact
-            1.0
-        };
-
-        // Update creature-level metrics
-        candidate.target_neuron_impact = impact;
-        let original = candidate.expected_creature_error_reduction;
-        candidate.expected_creature_error_reduction *= impact;
-        candidate.expected_creature_score_gain = candidate.expected_creature_error_reduction;
-
-        if verbose_enabled() && is_hidden {
-            eprintln!(
-                "[NEAT-AI-Discovery][verbose] Neuron candidate → {} impact {:.3}: \
-                {:.4}% → {:.4}%",
-                &candidate.target_neuron_uuid[..12.min(candidate.target_neuron_uuid.len())],
-                impact,
-                original * 100.0,
-                candidate.expected_creature_score_gain * 100.0
-            );
-        }
-    }
-
-    // Sort by expected creature score gain (highest first) - Issue #128
-    helpful_results.sort_by(|a, b| {
-        b.expected_creature_score_gain
-            .partial_cmp(&a.expected_creature_score_gain)
-            .unwrap_or(Ordering::Equal)
-    });
-
-    // Production experiment: pair "extreme" candidates with a conservative variant.
-    // Pass None for limit here - we'll truncate separately so that candidates_found
-    // correctly includes generated variants.
-    helpful_results = crate::analysis::utils::pair_extreme_candidates_with_conservative_variants(
-        helpful_results,
-        None, // No limit - truncate separately after capturing candidates_found
-    );
-
-    // Production guard rail (Dec 2025): only return candidates within sensible parameter ranges.
-    // This avoids wasting the evaluation budget on absurd bias/weight configurations.
-    helpful_results = crate::analysis::utils::filter_candidates_to_sensible_ranges(helpful_results);
-
-    // Deadline coverage (Jan 2026): when deadline-constrained, diversify within the top-K so
-    // repeated runs explore different high-quality candidates over time (helps with failure caches).
-    if input.analysis_deadline_ms.is_some() {
-        const DIVERSIFY_TOP_K: usize = 64;
-        shuffle_within_top_k(
-            helpful_results.as_mut_slice(),
-            input.random_seed,
-            "neuron:candidates:top_k",
-            DIVERSIFY_TOP_K,
-        );
-    }
-
-    // Track candidates_found AFTER pairing but BEFORE truncation.
-    // This ensures candidates_found >= candidates_returned always holds, which is
-    // the expected semantic for this metric pair ("found" >= "returned").
-    let candidates_found = helpful_results.len();
-
-    // Apply max_candidates limit (truncation)
-    if let Some(limit) = input.max_candidates {
-        helpful_results.truncate(limit);
-    }
-
-    // Track candidates_returned AFTER truncation
-    let candidates_returned = helpful_results.len();
-
-    let no_candidate_reasons = diagnostics.no_candidate_summaries();
-    diagnostics.emit_logs();
-
-    Ok(AnalyzeNeuronsResult {
-        helpful_neurons: helpful_results,
-        gpu_used,
-        no_candidate_reasons,
-        metadata: super::shared::NeuronAnalysisMetadata {
-            candidates_found,
-            candidates_returned,
-            timed_out: analysis_timed_out,
-            completed_focus_neurons: completed_count.load(std::sync::atomic::Ordering::Relaxed),
-            // Total focus neurons requested for this invocation (pre-filter).
-            //
-            // Note: `total_focus_count` is the post-filter eligible output-neuron count, which can
-            // differ from the requested focus list. We keep the "requested" semantics so callers
-            // can track long-run coverage consistently across early/normal return paths.
-            total_focus_neurons: original_focus_count,
-            timing: timing_collector.finalize(),
-            gpu_info: GpuAnalyzer::get_adapter_info(),
-        },
-    })
-}
-
-pub fn analyze_neurons(input: &AnalyzeNeuronsInput) -> Result<AnalyzeNeuronsResult> {
-    // Validate focus_neurons before expensive pre-loading
-    require_unique_focus(&input.focus_neurons, "Neuron analysis")?;
-
-    // Pre-load all records for faster analysis (1 scan vs ~2000 scans)
-    let cache = Arc::new(RecordCache::new_adaptive(&input.parquet_file)?);
-    analyze_neurons_with_cache(input, cache)
-}
-
-#[cfg(test)]
-mod tests {
-    // Unit tests live under `tests/` where possible, but remain unit tests so they can
-    // access private helpers without changing visibility.
-    include!(concat!(
-        env!("CARGO_MANIFEST_DIR"),
-        "/tests/unit/analysis_implementation.rs"
-    ));
-}
-// analyze_all has been moved to src/analysis/mod.rs
-
-/// Truncate synapse candidate buckets to a global cap, preserving ordering semantics.
-///
-/// Notes (7-Jan-2026):
-/// - Synapse analysis applies this during candidate assembly.
-/// - `analysis::analyze_all` may add coordinated candidates during post-processing (Issue #173),
-///   so it also reuses this helper to re-apply `maxSynapseCandidates` and keep output sizes stable.
-pub(crate) fn truncate_combined_synapse_candidate_sets(
-    helpful: Vec<CandidateSynapseJson>,
-    harmful: Vec<CandidateSynapseJson>,
-    coordinated: Vec<crate::CoordinatedStructuralCandidateJson>,
-    limit: usize,
-    diversify: bool,
-) -> (
-    Vec<CandidateSynapseJson>,
-    Vec<CandidateSynapseJson>,
-    Vec<crate::CoordinatedStructuralCandidateJson>,
-) {
-    if limit == 0 {
-        return (Vec::new(), Vec::new(), Vec::new());
-    }
-
-    // If we're already under the global cap, keep the caller's ordering exactly.
-    // This matters when upstream has intentionally shuffled within the top-K to diversify
-    // repeated deadline-limited runs (Jan 2026).
-    let total = helpful.len() + harmful.len() + coordinated.len();
-    if total <= limit {
-        return (helpful, harmful, coordinated);
-    }
-
-    // In diversified mode we intentionally preserve the per-bucket ordering (which may have been
-    // shuffled within the top-K) and select candidates in a round-robin fashion across buckets.
-    //
-    // This avoids "category starvation" when `max_candidates` is small: without this, one bucket
-    // with slightly higher expected gains can dominate the global sort and the other buckets may
-    // contribute zero candidates.
-    if diversify {
-        use std::collections::VecDeque;
-
-        let mut helpful_q: VecDeque<CandidateSynapseJson> = VecDeque::from(helpful);
-        let mut harmful_q: VecDeque<CandidateSynapseJson> = VecDeque::from(harmful);
-        let mut coordinated_q: VecDeque<crate::CoordinatedStructuralCandidateJson> =
-            VecDeque::from(coordinated);
-
-        let mut helpful_out = Vec::new();
-        let mut harmful_out = Vec::new();
-        let mut coordinated_out = Vec::new();
-
-        let mut returned = 0usize;
-        while returned < limit {
-            let mut progressed = false;
-
-            if let Some(c) = helpful_q.pop_front() {
-                helpful_out.push(c);
-                returned += 1;
-                progressed = true;
-                if returned >= limit {
-                    break;
-                }
-            }
-            if let Some(c) = harmful_q.pop_front() {
-                harmful_out.push(c);
-                returned += 1;
-                progressed = true;
-                if returned >= limit {
-                    break;
-                }
-            }
-            if let Some(c) = coordinated_q.pop_front() {
-                coordinated_out.push(c);
-                returned += 1;
-                progressed = true;
-                if returned >= limit {
-                    break;
-                }
-            }
-
-            if !progressed {
-                // All buckets are empty.
-                break;
-            }
-        }
-
-        return (helpful_out, harmful_out, coordinated_out);
-    }
-
-    enum Any {
-        Helpful(CandidateSynapseJson),
-        Harmful(CandidateSynapseJson),
-        Coordinated(crate::CoordinatedStructuralCandidateJson),
-    }
-
-    fn score(item: &Any) -> f32 {
-        match item {
-            Any::Helpful(c) => c.expected_creature_score_gain,
-            Any::Harmful(c) => c.expected_creature_score_gain,
-            Any::Coordinated(c) => c.expected_creature_score_gain,
-        }
-    }
-
-    let mut combined: Vec<Any> = Vec::with_capacity(total);
-    combined.extend(helpful.into_iter().map(Any::Helpful));
-    combined.extend(harmful.into_iter().map(Any::Harmful));
-    combined.extend(coordinated.into_iter().map(Any::Coordinated));
-
-    combined.sort_by(|a, b| {
-        score(b)
-            .partial_cmp(&score(a))
-            .unwrap_or(std::cmp::Ordering::Equal)
-    });
-    combined.truncate(limit);
-
-    let mut helpful_out = Vec::new();
-    let mut harmful_out = Vec::new();
-    let mut coordinated_out = Vec::new();
-    for item in combined {
-        match item {
-            Any::Helpful(c) => helpful_out.push(c),
-            Any::Harmful(c) => harmful_out.push(c),
-            Any::Coordinated(c) => coordinated_out.push(c),
-        }
-    }
-
-    (helpful_out, harmful_out, coordinated_out)
-}
 
 /// Internal implementation of synapse analysis with cache.
 /// This is called from the synapse module which owns the public API.
@@ -4496,15 +1800,38 @@ pub(crate) fn analyze_synapses_with_cache_impl(
 mod tests_synapses {
     use super::*;
     use crate::analysis::analyze_all;
+    use crate::analysis::analyze_neurons;
     use crate::analysis::analyze_synapses;
     use crate::analysis::samples::HelpfulStats;
     use crate::parquet_format::write_records_to_parquet;
-    use crate::{AnalyzeAllInput, CreatureJson, NeuronJson, SynapseJson};
+    use crate::{AnalyzeAllInput, AnalyzeNeuronsInput, CreatureJson, NeuronJson, SynapseJson};
     use std::sync::atomic::{AtomicUsize, Ordering as AtomicOrdering};
     use std::sync::{Arc, Barrier};
     use std::thread;
     use std::time::Duration;
     use tempfile::tempdir;
+
+    // Import functions moved to synapse.rs for tests (Issue #185)
+    use crate::analysis::synapse::{
+        build_samples, compute_activation_improvement_and_count,
+        compute_net_improvement_with_squash, compute_synapse_improvement_and_count,
+        compute_synapse_improvement_with_target_squash, count_improved_samples,
+        evaluate_relu_candidates_split, upsert_candidate,
+    };
+    // Import activation functions used by tests
+    use crate::analysis::activation::{
+        bipolar_activation, identity_activation, is_threshold_activation, logistic_activation,
+        tanh_activation,
+    };
+    // Import weights used by tests
+    use crate::analysis::weights::{calculate_optimal_bias, MAX_OUTGOING_WEIGHT};
+    // Import diagnostics types used by tests
+    use crate::analysis::diagnostics::{
+        filter_focus_targets_for_neuron_analysis, NeuronDiagnostics,
+    };
+    use crate::analysis::shared::NeuronNoCandidateReason;
+    // Import types used by tests
+    use crate::CandidateNeuronJson;
 
     /// Helper macro to skip tests that require GPU when no GPU is available.
     /// This allows tests to pass gracefully in CI environments without GPUs.
@@ -8937,6 +6264,12 @@ mod tests_optimal_outgoing_weight {
     use crate::analysis::samples::ReluStats;
     use anyhow::anyhow;
 
+    // Import types used by tests (Issue #185)
+    use crate::analysis::activation::{identity_activation, ActivationCandidateSpec};
+    use crate::analysis::gpu::GpuEvaluator;
+    use crate::analysis::synapse::evaluate_activation_candidate;
+    use crate::analysis::weights::MAX_OUTGOING_WEIGHT;
+
     struct AlwaysFailGpuEvaluator;
 
     impl GpuEvaluator for AlwaysFailGpuEvaluator {
@@ -9197,6 +6530,10 @@ mod tests_optimal_outgoing_weight {
 #[cfg(test)]
 mod tests_prediction_accuracy {
     use super::*;
+
+    // Import types used by tests (Issue #185)
+    use crate::analysis::synapse::compute_relu_improvement_and_count;
+    use crate::analysis::weights::MAX_OUTGOING_WEIGHT;
 
     /// Hard tanh activation for testing (same as production)
     fn test_hard_tanh(x: f32) -> f32 {
